% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/upgrade.R
\name{addSlot}
\alias{addSlot}
\title{Helper Function To Add A Slot to An Out-of-date S4 Object}
\usage{
addSlot(object, slot.name, slot.value)
}
\description{
Helper Function To Add A Slot to An Out-of-date S4 Object
}
\keyword{internal}

