% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{all.equal,conditionList-method}
\alias{all.equal,conditionList-method}
\title{Compare Lists of Conditions}
\usage{
\S4method{all.equal}{conditionList}(target, current, ...)
}
\arguments{
\item{target}{the list of conditions that we are matching against}

\item{current}{the list of conditions we are checking}
}
\value{
TRUE if the lists of conditions are equivalent, an character vector explaining
  why they are not otherwise
}
\description{
This is the default function for comparing conditions across the new and
reference runs of a test.  Comparison will fail if any of the conditions
are not all.equal.  Additionally, we check for the special "printed"
attribute on each conditions which indicate that the condition occurred
in the print/show methods as applied to the result of a call.
}
\keyword{internal}

