% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/conditions.R
\name{all.equal.condition}
\alias{all.equal.condition}
\title{Compare Conditions}
\usage{
\method{all}{equal.condition}(target, current, ...)
}
\arguments{
\item{target}{a condition}

\item{current}{another condition to compare}
}
\value{
TRUE if conditions match, character vector describing differences otherwise
}
\description{
Compare Conditions
}
\keyword{internal}

