% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.sub.R
\name{all.equal_}
\alias{all.equal_}
\title{Like all.equal but Returns "" If Not all.equal}
\usage{
\method{all}{equal_}(target, current, ...)
}
\arguments{
\item{target}{R object}

\item{current}{other R object to be compared to \code{target}}

\item{...}{arguments to pass to \code{\link{all.equal}}}
}
\value{
TRUE if \code{all.equal} returns TRUE, "" otherwise
}
\description{
Used as the default value comparison function since when values mismatch
we use \code{\link{diffObj}} which would make the text output from
\code{\link{all.equal}} somewhat redundant.
}

