% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ancestry_descend}
\alias{ancestry_descend}
\title{For Each ID Determines Generation}
\usage{
ancestry_descend(ids, par.ids, id, level = 0L)
}
\arguments{
\item{ids}{integer() the object ids}

\item{par.ids}{integer() the parents of each \code{ids}}

\item{id}{integer() the first parent}
}
\value{
matrix containing ids and corresponding generation for the ids
}
\description{
For Each ID Determines Generation
}
\keyword{internal}

