% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{append,factor,factor-method}
\alias{append,factor,factor-method}
\title{Append Factors}
\usage{
\S4method{append}{factor,factor}(x, values, after = length(x))
}
\description{
Note this is not related to \code{`\link{append,unitizerList,ANY-method}`}
except in as much as it is the same generic, so it just got thrown in here.
}
\keyword{internal}

