% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{append,unitizerList,ANY-method}
\alias{append,unitizerList,ANY-method}
\title{Append To a \code{`\link{unitizerList-class}`} Object}
\usage{
\S4method{append}{unitizerList,ANY}(x, values, after = length(x))
}
\arguments{
\item{x}{the object to append to}

\item{values}{the object to append}

\item{after}{a subscript, after which the values are to be appended.}
}
\description{
\code{`values`} is coerced to list or expression depending on
type of \code{`x`} \code{`.items`} slot.
}
\details{
The resulting object must pass the validity method for \code{`x`}.
}
\keyword{internal}

