% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{as.character,testorItemTestsErrors-method}
\alias{as.character,testorItemTestsErrors-method}
\title{Convert An Error Into Character}
\usage{
\S4method{as.character}{testorItemTestsErrors}(x, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Convert An Error Into Character
}
\keyword{internal}

