% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/browse.struct.R
\docType{methods}
\name{as.character,unitizerBrowse-method}
\alias{as.character,unitizerBrowse-method}
\title{Create a Text Representation of an Object}
\usage{
\S4method{as.character}{unitizerBrowse}(x, width = 0L, ...)
}
\arguments{
\item{width}{how many characters to display max per line, use 0L to use the
terminal window width as determined by \code{`getOption("width")`}; note
this is a guideline, if you pass numbers that lead to too narrow renderings
it will be ignored.}

\item{...}{not used}

\item{object}{object to render}
}
\value{
character vector, one element per line, for use with e.g.
  \code{`cat(x, sep=\n`)}
}
\description{
Create a Text Representation of an Object
}
\keyword{internal}

