% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\docType{methods}
\name{as.character,unitizerChanges-method}
\alias{as.character,unitizerChanges-method}
\title{Print Out A Summary Of the Changes}
\usage{
\S4method{as.character}{unitizerChanges}(x, width = getOption("width"),
  ...)
}
\description{
Print Out A Summary Of the Changes
}
\keyword{internal}
