% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/item.sub.R
\docType{methods}
\name{as.character,unitizerItemTestsErrors-method}
\alias{as.character,unitizerItemTestsErrors-method}
\title{Convert An Error Into Character}
\usage{
\S4method{as.character}{unitizerItemTestsErrors}(x, ...)
}
\description{
Convert An Error Into Character
}
\keyword{internal}

