% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{as.list,testorList-method}
\alias{as.list,testorList-method}
\title{Coerce to list by returning items}
\usage{
\S4method{as.list}{testorList}(x, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{objects, possibly named.}
}
\description{
Coerce to list by returning items
}
\keyword{internal}

