% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load.R
\name{as.store_id_chr}
\alias{as.store_id_chr}
\title{Convert Store ID to Character}
\usage{
as.store_id_chr(x)
}
\description{
For display purposes only since path is relativized.
}
\details{
If not possible make up a name
}
\keyword{internal}

