% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/asciiml.R
\name{asciiml}
\alias{asciiml}
\title{Text Representations of HTML Objects}
\description{
Functions defined here should ultimately be extended into a more
comprehensive stand alone library for ASCII structured objects (banners,
tables, etc.).
}
\details{
Blah b
}
\keyword{internal}

