% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load.R
\name{best_store_name}
\alias{best_file_name}
\alias{best_store_name}
\title{Get Most Intuitive Name for Store}
\usage{
best_store_name(store.id, test.file)

best_file_name(store.id, test.file)
}
\arguments{
\item{store.id}{a \code{unitizer} store id}

\item{test.file}{the location of the R test file}
}
\value{
character(1L)
}
\description{
Based on data from \code{store.id} and \code{test.file}
}
\keyword{internal}

