% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/browse.struct.R
\name{browsePrep}
\alias{browsePrep}
\title{Prepares a \code{`unitizer`} Object for Review}
\usage{
browsePrep(x, mode, ...)
}
\description{
Mainly, splits up the tests by section and subsection and creates an indexing
structure to keep track of what tests are in which section/subsection.  This
simplifies implementation of non-linear navigation through the tests.
}
\keyword{internal}

