% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{browseUnitizer,unitizer,unitizerBrowse-method}
\alias{browseUnitizer,unitizer,unitizerBrowse-method}
\title{Browse unitizer}
\usage{
\S4method{browseUnitizer}{unitizer,unitizerBrowse}(x, y, prompt.on.quit,
  force.update, ...)
}
\arguments{
\item{x}{the object to browse}

\item{y}{the derivative unitizerBrowse object of x; this needs to be passed in
as an argument because the logic for generating it is different depending on
whether we are using `unitize` or `review`.}

\item{prompt.on.quit}{whether to prompt for review even if there are no changes}

\item{a}{unitizer if the unitizer was modified, FALSE otherwise}
}
\description{
Here we are reviewing all the tests in the unitizer under three different lenses
\enumerate{
  \item tests that don't match the stored reference tests
  \item tests that don't exist in the reference tests
  \item tests that exist in the reference tests but no the new file
  \item tests that passed (these are omitted )
}
Because a lot of the logic for browsing these three types of situations is
shared, that logic has been split off into \code{`\link{reviewNext,unitizerBrowse-method}`}.
The key is that that function will return the items that are supposed to be
stored in the unitizer.  These items will either be new or reference ones
based on user decisions.
}
\details{
Unfortunately, in order to be able to use the same logic for tasks that are
not quite the same, a bit of contortion was needed.  In particular, the
user is always asked to input either Y, N, or Q, but the corresponding output
from \code{`\link{reviewNext,unitizerBrowse-method}`} is very different
depending on what situation we're dealing with.

One important point is that by default the user input is defined as N.  In
all cases N means no change to the store, though again the interpretation is
different depending on the situation.  For example, if we add a test to the
test script, N means don't add it to the store.  If we remove a test, N means
keep it in the store.
}
\keyword{internal}

