% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/list.R
\docType{methods}
\name{c,unitizerList-method}
\alias{c,unitizerList-method}
\title{Concatenate to a \code{`\link{unitizerList-class}`}}
\usage{
\S4method{c}{unitizerList}(x, ..., recursive = FALSE)
}
\description{
Concatenate to a \code{`\link{unitizerList-class}`}
}
\keyword{internal}

