% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cap_first}
\alias{cap_first}
\alias{decap_first}
\title{Captalizes or Decapitalizes First Letter}
\usage{
cap_first(x)
}
\arguments{
\item{x}{character}
}
\value{
character
}
\description{
Captalizes or Decapitalizes First Letter
}
\keyword{internal}

