% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{capture_output}
\alias{capture_output}
\alias{print.captured_output}
\title{Capture Both StdOut and StdErr}
\usage{
capture_output(expr, env = parent.frame())

\method{print}{captured_output}(x, ...)
}
\arguments{
\item{env}{an environment to evaluate them in}

\item{a}{quoted to evaluate}
}
\value{
a list with stdout and stderr captured separately, classed as
  "captured_output"
}
\description{
Will sink both "output" and "message" streams without checking whether they
are already sunk, and will unsink them the same way.
}
\keyword{internal}
