% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unitize.core.R
\name{check_call_stack}
\alias{check_call_stack}
\title{Check Not Running in Undesirable Environments}
\usage{
check_call_stack()
}
\description{
Make sure not running inside withCallingHandlers / withRestarts / tryCatch
or other potential issues; of course this isn't foolproof if someone is using
a variation on those functions, but also not the end of the world if it isn't
caught
}
\keyword{internal}

