% Generated by roxygen2 (4.0.1): do not edit by hand
\name{comm_reset}
\alias{comm_reset}
\title{Utility Function to Reset Comments}
\usage{
comm_reset(x)
}
\description{
Required due to bizarre behavior (bug?) where some expression attributes
appear to have reference like behavior even when they are re-generated
from scratch from a text expression (wtf, really).
}

