\name{compare_conditions}
\alias{compare_conditions}
\title{Compare Lists of Conditions}
\usage{
compare_conditions(target, current)
}
\arguments{
  \item{target}{the list of conditions that we are matching
  against}

  \item{current}{the list of conditions we are checking}
}
\value{
TRUE if the lists of conditions are equivalent, an
character vector explaining why they are not otherwise
}
\description{
Compare Lists of Conditions
}

