% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unitize.core.R
\name{confirm_quit}
\alias{confirm_quit}
\title{Prompt to Quit if Enough Time Spent on Evaluation}
\usage{
confirm_quit(unitizers)
}
\description{
Prompt to Quit if Enough Time Spent on Evaluation
}
\keyword{internal}

