% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/text.R
\name{deparse_fun}
\alias{deparse_fun}
\title{Returns a Character Function Name From A Language Object}
\usage{
deparse_fun(x)
}
\arguments{
\item{x}{a call or a symbol}
}
\value{
character 1 length if a function name, NA if an anonymous function, or
  character(0L) if neither
}
\description{
Note this doesn't really try to check too hard whether the \code{`x`} is
indeed a function.
}
\keyword{internal}

