% Generated by roxygen2 (4.0.2): do not edit by hand
\name{deparse_mixed}
\alias{deparse_mixed}
\title{Special Deparse}
\usage{
deparse_mixed(expr, width.cutoff = 500L, control = "all", ...)
}
\description{
Required to deal with language objects that contain non-language objects
that have attributes.
}
\details{
Not completely fool proof since you can probably created an object that nests
call and non-call stuff repeatedly that would confuse this thing.
}
\keyword{internal}

