% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/text.R
\name{deparse_peek}
\alias{deparse_peek}
\title{Deparse, but only provide first X characters}
\usage{
deparse_peek(expr, len, width = 500L)
}
\arguments{
\item{expr}{a language object}

\item{len}{int a one length integer noting how many characters we want}

\item{width}{passed on to}
}
\description{
Deparse, but only provide first X characters
}
\keyword{internal}

