% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/text.R
\name{deparse_prompt}
\alias{deparse_prompt}
\title{Deparse, But Make It Look Like It Would On Prompt}
\usage{
deparse_prompt(expr)
}
\arguments{
\item{expr}{an expression or call}
}
\value{
character vector
}
\description{
Deparse, But Make It Look Like It Would On Prompt
}
\keyword{internal}

