% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{diff_internal}
\alias{diff_internal}
\title{Implements the diff_* functions}
\usage{
diff_internal(obj.rem, obj.add, context = NULL, mode = "auto",
  obj.rem.name = deparse(substitute(obj.rem))[[1L]],
  obj.add.name = deparse(substitute(obj.add))[[1L]],
  width = getOption("width"), file = stdout(), frame = parent.frame())
}
\arguments{
\item{width}{at what width to wrap output}

\item{file}{whether to show to stdout or stderr}

\item{frame}{what frame to capture in, relevant mostly if looking for a print
method}
}
\description{
Implements the diff_* functions
}
\keyword{internal}

