% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{diff_obj}
\alias{diff_obj}
\alias{diff_print}
\alias{diff_str}
\title{Show Diffs Between The Display Values of Two Objects}
\usage{
diff_obj(target, current, context = NULL)

diff_print(target, current, context = NULL)

diff_str(target, current, context = NULL, max.level = 10)
}
\arguments{
\item{target}{the reference object}

\item{context}{2 length integer vector representing how many lines of context
are shown on either side of differences.  The first value is the maximum
before we start trimming output.  The second value is the maximum to be
shown before we start trimming.  We will always attempt to show as much as
\code{2 * context + 1} lines of output so context may not be centered if
objects display as less than \code{2 * context + 1} lines.}

\item{max.level}{integer(1L) up to how many levels to try running \code{str};
\code{str} is run repeatedly starting with \code{max.level=1} and then
increasing \code{max.level} until a difference appears or the
\code{max.level} specified here is reached.  If the value is reached then
will let \code{str} run with \code{max.level} unspecified.  This is
designed to produce the most compact screen output possible that shows the
differences between objects, though obviously it comes at a performance
cost; set to 0 to disable}

\item{new}{the object being compared to \code{target}}
}
\value{
character, invisibly, the text representation of the diff
}
\description{
Designed to highlight at a glance the \bold{display} differences between
two objects.  Note that the difference shown or reported by these functions
may not be the totality of the differences between objects since display
methods may not display all differences.
}
\details{
\itemize{
  \item \code{diff_print} shows the differences in the \code{print} or
    \code{show} screen output of the two objects
  \item \code{diff_str} shows the differences in the \code{str} screen output
    of the two objects; will show as few recursive levels as needed to show
    at least one difference (see \code{max.level})
  \item \code{diff_obj} picks between \code{diff_print} and \code{diff_str}
    depending on which one it thinks will provide the most useful diff
}
}

