% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.compare.R
\name{diff_state}
\alias{diff_state}
\title{Display State Differences Between New and Reference Tests}
\usage{
diff_state(target = NULL, current = NULL, width = getOption("width"),
  file = stdout())
}
\arguments{
\item{target}{unitizerGlobalState object}

\item{current}{unitizerGlobalState object}

\item{width}{how many characters wide the display should be}

\item{file}{connection to output to}
}
\description{
Intended to be called primarily from \code{unitizer} prompt where it can
find the \code{.NEW} and \code{.REF} objects; interface params provided for
testing.
}
\examples{
\dontrun{
## type at `unitizer` prompt
diff_state()
}
}

