% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/browse.struct.R
\docType{methods}
\name{done,unitizerBrowse-method}
\alias{done,unitizerBrowse-method}
\title{Indicate Whether to Exit Review Loop}
\usage{
\S4method{done}{unitizerBrowse}(x, ...)
}
\description{
Indicate Whether to Exit Review Loop
}
\keyword{internal}

