\name{env_name}
\alias{env_name}
\title{Gets Environment Name / Memory Code}
\usage{
env_name(env)
}
\arguments{
  \item{env}{an environemnt}
}
\value{
character 1 length
}
\description{
Captures the name that \code{`\link{print.default}`}
displays when one prints and environment
}
\keyword{internal}

