% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eval_user_exp}
\alias{eval_user_exp}
\title{Utility function to evaluate user expressions}
\usage{
eval_user_exp(unitizerUSEREXP, env)
}
\arguments{
\item{unitizerUSEREXP}{an expression to evaluate}

\item{env}{environment the environment to evaluate the expression in}
}
\value{
TBD
}
\description{
Utility function to evaluate user expressions
}
\seealso{
exec, unitizer_prompt
}
\keyword{internal}

