% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/exec.R
\docType{methods}
\name{exec,ANY-method}
\alias{exec,ANY-method}
\title{Manages Test Expression Evaluation}
\usage{
\S4method{exec}{ANY}(x, test.env, capt.cons)
}
\arguments{
\item{test.env}{the environment to evaluate the \code{test} in}

\item{test}{the call to test}
}
\value{
a \code{\link{unitizerItem-class}} object
}
\description{
Runs test, captures value, stdout, stderr, conditions, etc.
}
\keyword{internal}

