% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prompt.R
\name{exit_fun}
\alias{exit_fun}
\title{An Exit Fun For Prompts Expecting a Specific Selection}
\usage{
exit_fun(y, env, valid.vals)
}
\arguments{
\item{valid.vals}{vector to check user input against}
}
\description{
Intended for use solely with \code{\link{unitizer_prompt}}
}
\keyword{internal}

