% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.sub.R
\docType{methods}
\name{$,unitizerItemTestsErrorsDiffs-method}
\alias{$,unitizerItemTestsErrorsDiffs-method}
\alias{[[,unitizerItemTestsErrorsDiffs,ANY-method}
\title{Subsetting Methods for unitizerItemTestsErrorsDiffs objects}
\usage{
\S4method{$}{unitizerItemTestsErrorsDiffs}(x, name)

\S4method{[[}{unitizerItemTestsErrorsDiffs,ANY}(x, i, j, ...,
  exact = TRUE)
}
\description{
Subsetting Methods for unitizerItemTestsErrorsDiffs objects
}
\keyword{internal}
