% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/browse.struct.R
\name{extractItems}
\alias{extractItems}
\title{Pull out items from unitizerBrowse objects}
\usage{
extractItems(x, ...)
}
\description{
Pull out items from unitizerBrowse objects
}
\keyword{internal}

