% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fastlm_dir}
\alias{fastlm_dir}
\alias{prompt_to_proceed}
\title{Functions To Assist With Demo}
\usage{
fastlm_dir(version)

prompt_to_proceed()
}
\arguments{
\item{version}{a number in 0:2}
}
\value{
character(1L)
}
\description{
\code{`fastlm_dir`} returns the directory that the \code{`unitizer.fastlm`}
package is in, and \code{`prompt_to_proceed`} stops demo evaluation until
user presses ENTER.
}
\note{
these functions are not for use outside of the unitizer demo
}

