% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{filename_to_storeid}
\alias{filename_to_storeid}
\title{Create a Store ID from a Test File Name}
\usage{
filename_to_storeid(x)
}
\arguments{
\item{x}{character(1L) file name ending in .r or .R}
}
\value{
store id name, or NULL if \code{x} doesn't meet expectations
}
\description{
Create a Store ID from a Test File Name
}
\examples{
filename_to_storeid(file.path("tests", "unitizer", "foo.R"))
filename_to_storeid(file.path("tests", "unitizer", "boo.r"))
# does not end in [rR]
filename_to_storeid(file.path("tests", "unitizer", "boo"))
}
