% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{flattenUntz}
\alias{flattenUntz}
\title{Reduce S4 objects Into Lists}
\usage{
flattenUntz(x, ...)
}
\description{
This is particularly useful with "list" type S4 objects, and relates loosely
to the subsetting functions defined for \code{unitizerBrowse} objects.
}
\details{
Currently we only define a method for \code{unitizerItems-class} objects
}
\keyword{internal}
