% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{funs.ignore}
\alias{funs.ignore}
\title{Functions To Ignore}
\format{\preformatted{List of 4
 $ :.Primitive("<-") 
 $ :function (package, help, pos = 2, lib.loc = NULL, character.only = FALSE, 
    logical.return = FALSE, warn.conflicts = TRUE, quietly = FALSE, verbose = getOption("verbose"))  
 $ :.Primitive("=") 
 $ :function (seed, kind = NULL, normal.kind = NULL)  
}}
\usage{
funs.ignore
}
\arguments{
\item{x}{the reference function, if is traced then y must be identical}

\item{y}{the current function, if \code{`x`} is not traced and \code{`y`}
is traced, will compare using \code{`y@original`} instead of \code{`y`}}
}
\description{
Ignored functions are not considered tests if they are called from
the top level.
}
\details{
Also, provide a function to compare functions even when traced.
}
\keyword{internal}

