% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/conditions.R
\name{get_condition_type}
\alias{get_condition_type}
\title{Extracts Condition Type From Condition Classes}
\usage{
get_condition_type(x)
}
\arguments{
\item{x}{a condition}
}
\value{
character 1 length the type of condition
}
\description{
Type (e.g. Error, Warning), is taken to be the second to last class.
}
\keyword{internal}

