% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_trace}
\alias{get_trace}
\title{Collect the Call Stack And Clean-up}
\usage{
get_trace(trace.base, trace.new, printed, print.type, exp)
}
\arguments{
\item{trace.base}{starting point for what we care about in the trace, as
produced by \code{sys.calls}}

\item{trace.new}{the trace within the condition handler, as produced by
\code{sys.calls}}

\item{print.type}{character(1L) one of "print", "show", or ""}

\item{exp}{the expression to sub in to the print/show statements if we passed
eval}

\item{passsed.eval}{whether the evaluatation succeeded in the first step (see
details)}
}
\description{
Only intended for use within \code{eval_user_exp}, will clean up the result
from two different \code{sys.calls} calls to extract the calls that a
trace would show on error.
}
\details{
How much of the stack is used is affected by the \code{printed}
argument because if something didn't pass evaluation, it means the error
occurred within \code{withVisible} which in this setup means we need to
remove two additional levels.

Relies on calls being evaluated in a very particular environment.
}
\seealso{
set_trace
}
\keyword{internal}

