% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unitizer.R
\docType{methods}
\name{getTarget,unitizer-method}
\alias{getTarget,unitizer-method}
\title{Create A Human Readable Names for a \code{unitizer}}
\usage{
\S4method{getTarget}{unitizer}(object, ...)
}
\arguments{
\item{object}{a unitizer}
}
\value{
character(1L) a descriptive name
}
\description{
Create A Human Readable Names for a \code{unitizer}
}
\keyword{internal}

