% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/asciiml.R
\name{header}
\alias{H1,}
\alias{H2,}
\alias{H3}
\alias{header}
\title{Create Header Objects}
\usage{
header(x, level)
}
\arguments{
\item{x}{1 length character vector to turn into a header}

\item{level}{1 length integer, what level to make a header}
}
\value{
header object
}
\description{
Header objects are 1 length character vectors that are printed with text
formatting that highlight their "headerness".
}
\seealso{
\code{`\link{print.header}`}
}
\keyword{internal}

