% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/item.sub.R
\docType{methods}
\name{initialize,unitizerItemTestsFuns-method}
\alias{initialize,unitizerItemTestsFuns-method}
\title{Ensures Functions are In Correct Format}
\usage{
\S4method{initialize}{unitizerItemTestsFuns}(.Object, ...)
}
\description{
Also, allows the user to specify functions directly instead of having
to instantiate \code{`\link{unitizerItemTestFun-class}`} for each function.
Finally, recovers the deparsed passed function name.
}
\keyword{internal}

