% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls.R
\name{invalidateLs}
\alias{invalidateLs}
\title{Clears ls Info and Marks as Invalid}
\usage{
invalidateLs(x, ...)
}
\description{
Useful when tests envs are repaired, or if we're looking at an ignored
test
}
\keyword{internal}
