% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/is.R
\name{is.chr1plain}
\alias{is.chr1plain}
\title{Test for plain characterness}
\usage{
is.chr1plain(x)
}
\description{
Used primarily for assessing whether a store id should use default mechanism
or should be coerced to character
}
\keyword{internal}

