% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zeroenv.R
\name{is.loaded_package}
\alias{is.loaded_package}
\title{Check Whether a Package Is Loaded}
\usage{
is.loaded_package(pkg.name)
}
\arguments{
\item{pkg.name}{character(1L) must be in format "package:pkgname"}
}
\value{
TRUE if it is a loaded package
}
\description{
A package is considered loaded if it is in the search path and there is a
namespace loaded with the same name as the package
}
\keyword{internal}

