% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/is.R
\name{is.package_version}
\alias{is.package_version}
\title{Confirm Object is In \code{package_version} form}
\usage{
is.package_version(x)
}
\description{
Confirm Object is In \code{package_version} form
}
\keyword{internal}

