% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/is.R
\name{is.chr1plain}
\alias{is.TF}
\alias{is.chr1}
\alias{is.chr1plain}
\alias{is.int.pos.2L}
\title{Test for plain characterness}
\usage{
is.chr1plain(x)

is.chr1(x)

is.TF(x)

is.int.pos.2L(x)
}
\arguments{
\item{x}{object to test}
}
\description{
Test for common scalar cases that we run into ALL THE TIME!!!
}
\keyword{internal}

