% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.R
\name{is_package_dir}
\alias{is_package_dir}
\title{Check Whether a Directory Likey Contains An R Package}
\usage{
is_package_dir(name, has.tests = FALSE)
}
\arguments{
\item{name}{a directory to check for package-ness}

\item{has.tests}{whether to require that the package have tests to qualify}
}
\value{
TRUE if criteria met, character vector explaining first failure
  otherwise
}
\description{
Approximate check based on DESCRIPTION file and directory structure.
}
\keyword{internal}

