% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/translate.R
\name{is_testthat_attached}
\alias{is_testthat_attached}
\title{Confirm that `testthat` Is Attached}
\usage{
is_testthat_attached()
}
\description{
Would normally do this via NAMESPACE, but do not want to introduce an
explicit dependency to \code{testthat} since it is only required for the
conversion to \code{unitizer}
}

