% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.R
\name{is_unitizer_dir}
\alias{is_unitizer_dir}
\title{Check Whether a Directory as a Unitizer Data Directory}
\usage{
is_unitizer_dir(dir)
}
\arguments{
\item{dir}{character(1L) directory to check}
}
\value{
logical(1L)
}
\description{
Just checks that it \emph{could} be a data directory, the test ultimately is
to attempt a \code{\link{get_unitizer}} call and see if we actually resurrect
a working \code{unitizer}
}
\keyword{internal}

