% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/item.R
\name{isValid}
\alias{isValid}
\title{Used to run "validity" methods}
\usage{
isValid(object, ...)
}
\arguments{
\item{object}{object to validate}

\item{...}{additional arguments}
}
\value{
TRUE on success, character vector explaining failure otherwise
}
\description{
This is to avoid having to automatically run validity on object
instantiation as is the case with actual built-in validity objects.  We can
create a bunch of objects, and then only run validity when
}
\keyword{internal}

