% Generated by roxygen2 (4.0.1): do not edit by hand
\name{itemType}
\alias{itemType}
\alias{itemType,unitizerItem-method,}
\alias{itemType<-,unitizerItem-method,}
\alias{itemsType<-,unitizerItems-method}
\alias{itemsitemType,unitizerItems-method,}
\title{Methods to Track Whether a \code{`\link{unitizerItem-class}`} Object is New Or Reference}
\usage{
itemType(x, ...)
}
\description{
Necessitated due to the awkward structure around \code{`\link{reviewNext,unitizerBrowse-method}`},
where the only return value is a \code{`\link{unitizerItems-class}`} object and there is
no easy way to tell which objects have been kept from reference vs which ones are
new.
}
\keyword{internal}

