% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{length,unitizer-method}
\alias{length,unitizer-method}
\title{Compute Length of a \code{\link{unitizer-class}} Object}
\usage{
\S4method{length}{unitizer}(x)
}
\description{
Compute Length of a \code{\link{unitizer-class}} Object
}
\keyword{internal}

