% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{length,unitizerChanges-method}
\alias{length,unitizerChanges-method}
\title{Return Sum of Total Changes}
\usage{
\S4method{length}{unitizerChanges}(x)
}
\arguments{
\item{x}{an \R object.  For replacement, a vector or factor.}
}
\description{
Return Sum of Total Changes
}
\keyword{internal}

