% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{length,unitizerList-method}
\alias{length,unitizerList-method}
\title{Compute Number of Items in \code{`\link{unitizerList-class}`}}
\usage{
\S4method{length}{unitizerList}(x)
}
\description{
Compute Number of Items in \code{`\link{unitizerList-class}`}
}
\keyword{internal}

