% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{length,unitizerSection-method}
\alias{length,unitizerSection-method}
\title{Compute Length of a \code{`\link{unitizerSection-class}`}}
\usage{
\S4method{length}{unitizerSection}(x)
}
\arguments{
\item{x}{an \R object.  For replacement, a vector or factor.}
}
\description{
Compute Length of a \code{`\link{unitizerSection-class}`}
}

