% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load.R
\name{load_unitizer}
\alias{load_unitizer}
\alias{store_unitizer}
\title{Retrieve Unitizer}
\usage{
load_unitizer(store.id, par.frame)

store_unitizer(unitizer, store.id, wd)
}
\arguments{
\item{store.id}{anything for which there is a defined \code{`\link{get_unitizer}`}
method; by default should be the path to a unitizer; if \code{`\link{get_unitizer}`}
returns \code{`FALSE`} then this will create a new unitizer}

\item{par.frame}{the environment to use as the parent frame for the \code{`unitizer`}}
}
\value{
a \code{`unitizer`} object, or anything, in which case the calling
  code should exit
}
\description{
If this errors, calling function should abort
}
\keyword{internal}

