% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load.R
\name{load_unitizers}
\alias{load_unitizers}
\alias{store_unitizer}
\title{Store Retrieve Unitizer}
\usage{
load_unitizers(store.ids, test.files, par.frame, interactive.mode, mode,
  force.upgrade = FALSE)

store_unitizer(unitizer)
}
\arguments{
\item{par.frame}{the environment to use as the parent frame for the \code{unitizer}}

\item{force.upgrade}{whether to allow upgrades in non-interactive mode, for
testing purposes}

\item{unitizer}{a \code{\link{unitizer-class}} object}

\item{store.id}{anything for which there is a defined \code{`\link{get_unitizer}`}
method; by default should be the path to a unitizer; if \code{`\link{get_unitizer}`}
returns \code{`FALSE`} then this will create a new unitizer}

\item{test.file}{the R file associated with the store id}
}
\value{
a \code{unitizer} object, or anything, in which case the calling
  code should exit
}
\description{
If this errors, calling function should abort
}
\keyword{internal}

