% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{mock_item}
\alias{mock_item}
\title{Generates a Dummy Item For Use in Examples}
\usage{
mock_item()
}
\value{
unitizerItem object
}
\description{
The only purpose of this function is to create a \code{unitizerItem} for use
by examples.
}
