% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load.R
\name{norm_store_id}
\alias{norm_file}
\alias{norm_store_id}
\title{Get A Store ID in Full Path Format}
\usage{
norm_store_id(x)

norm_file(x)
}
\description{
Loosely related to \code{getTarget,unitizer-method} and
\code{getName,unitizer-method} although these are not trying to convert to
character or check anything, just trying to normalize if possible.
}
\details{
Relevant for default ids
}
\keyword{internal}

