% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options_zero}
\alias{options_zero}
\alias{options_update}
\alias{validate_options}
\title{Set Options to Initial Zero State}
\usage{
options_zero(base = merge_lists(getOption("unitizer.opts.init.base"),
  getOption("unitizer.opts.init")),
  as.is = union(getOption("unitizer.opts.asis.base"),
  getOption("unitizer.opts.asis")))

options_update(tar.opts)

validate_options(opts.to.validate, test.files = NULL)
}
\description{
This makes sure to unset options not present in target.
}
\keyword{internal}
