% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/options.R
\name{options_update}
\alias{options_update}
\title{Set Options to Specified State}
\usage{
options_update(tar.opts)
}
\description{
This makes sure to unset options not present in target.
}
\keyword{internal}

