% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/options.R
\name{options_zero}
\alias{options_zero}
\title{Set Options to Initial Zero State}
\usage{
options_zero(base = getOption("unitizer.opts.base"),
  as.is = getOption("unitizer.opts.asis"))
}
\description{
Set Options to Initial Zero State
}
\keyword{internal}

