% Generated by roxygen2 (4.0.2): do not edit by hand
\name{over_print}
\alias{over_print}
\title{Over-write a Line}
\usage{
over_print(x, min.width = 30L, max.width = getOption("width"))
}
\arguments{
\item{x}{character(1L)}

\item{min.width}{integer(1L) minimum character width to print to}

\item{max.width}{integer(1L) max width to print to}
}
\value{
NULL used only for side effect of cating ot screen
}
\description{
Over-write a Line
}
\keyword{internal}

