% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{pack.env}
\alias{objects.attached}
\alias{pack.env}
\alias{zero.env.par}
\title{Objects used to Track What Environment to Use as Parent to Zero Env}
\format{\preformatted{<environment: 0x7fe7a65018e8> 
}}
\usage{
pack.env
}
\description{
Note \code{`"package:unitizer"`} is not actually detached but is included in
the list so that it is replaced in the same position on the search path
}
\details{
Packages are stored as character values, other objects are stored directly.

Probably should be an RC object instead of an environment.
}
\keyword{internal}

