% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_dat_get}
\alias{parse_dat_get}
\title{Handle the issues with needing to run parse twice due to weird getParseData
output}
\usage{
parse_dat_get(file, text)
}
\description{
Handle the issues with needing to run parse twice due to weird getParseData
output
}
\keyword{internal}

