% Generated by roxygen2 (4.0.2): do not edit by hand
\name{passed}
\alias{passed}
\title{Determine if a \code{`unitizer`} Passed Based On Summary}
\usage{
passed(object, ...)
}
\arguments{
\item{object}{object to test for passing}
}
\value{
logical(1L)
}
\description{
Determine if a \code{`unitizer`} Passed Based On Summary
}
\keyword{internal}

