% Generated by roxygen2 (4.0.1): do not edit by hand
\name{path_clean}
\alias{path_clean}
\title{Cleans a Path to be In Standard Format}
\usage{
path_clean(path)
}
\description{
Uses \code{`\link{basedir}`} to convert paths on windows machines with back
slasshes to forward slash based names, and then removed excess forward
slashes.
}

