% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.R
\name{pick_one}
\alias{pick_one}
\title{Utility Fun to Poll User Input}
\usage{
pick_one(x, prompt = "Select Item Number:\\n")
}
\arguments{
\item{x}{a character vector}
}
\value{
integer(1L) selected item or 0L if invalid user input
}
\description{
Utility Fun to Poll User Input
}
\keyword{internal}

