% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{+,testor,testorItems-method}
\alias{+,testor,testorItems-method}
\title{Adds \code{`\link{testorItems-class}`} objects to Testor}
\usage{
\S4method{+}{testor,testorItems}(e1, e2)
}
\description{
Any added \code{`\link{testorItems-class}`} objects are treated as
reference items.  The only way to add new items is by adding each
item individually with \code{`\link{+,testor,testorItem-method}`}.
}

