% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{+,testor,testorItemTestsErrors-method}
\alias{+,testor,testorItemTestsErrors-method}
\title{Add Test Errors to \code{`\link{testor-class}`}}
\usage{
\S4method{+}{testor,testorItemTestsErrors}(e1, e2)
}
\description{
Add Test Errors to \code{`\link{testor-class}`}
}
\keyword{internal}

