% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{+,testor,testorSection-method}
\alias{+,testor,testorSection-method}
\title{Add a \code{`\link{testorSection-class}`} to a \code{`\link{testor-class}`}}
\usage{
\S4method{+}{testor,testorSection}(e1, e2)
}
\description{
Registers the section, and the mapping of items to section.
}
\keyword{internal}

