% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{+,testor,testorTestsOrExpression-method}
\alias{+,testor,testorTestsOrExpression-method}
\title{Adds Expressions to Testor}
\usage{
\S4method{+}{testor,testorTestsOrExpression}(e1, e2)
}
\description{
Expressions can be added as \code{`\link{testorTests-class}`} object
or a straight up expression, though in most cases it should be the
latter.
}
\note{
you can only do this once for a `testor`.
}

