% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{+,testorItems,testorItemOrNULL-method}
\alias{+,testorItems,testorItemOrNULL-method}
\title{Add a \code{`\link{testorItem-class}`} to a \code{`\link{testorItems-class}`}}
\usage{
\S4method{+}{testorItems,testorItemOrNULL}(e1, e2)
}
\description{
Add a \code{`\link{testorItem-class}`} to a \code{`\link{testorItems-class}`}
}
\keyword{internal}

