% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{+,unitizer,unitizerItem-method}
\alias{+,unitizer,unitizerItem-method}
\title{Adds \code{`\link{unitizerItem-class}`} to \code{`\link{unitizer-class}`}}
\usage{
\S4method{+}{unitizer,unitizerItem}(e1, e2)
}
\description{
All tests are run on addition, and mapping information between reference and
new tests is also recored.
}
\seealso{
\code{`\link{registerItem,unitizer,unitizerItem-method}`}
}
\keyword{internal}

