% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unitizer.add.R
\docType{methods}
\name{+,unitizer,unitizerItemTestsErrors-method}
\alias{+,unitizer,unitizerItemTestsErrors-method}
\title{Add Test Errors to \code{`\link{unitizer-class}`}}
\usage{
\S4method{+}{unitizer,unitizerItemTestsErrors}(e1, e2)
}
\description{
Add Test Errors to \code{`\link{unitizer-class}`}
}
\keyword{internal}

