% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{+,unitizer,unitizerTestsOrExpression-method}
\alias{+,unitizer,unitizerTestsOrExpression-method}
\title{Adds Expressions to unitizer}
\usage{
\S4method{+}{unitizer,unitizerTestsOrExpression}(e1, e2)
}
\description{
Expressions can be added as \code{`\link{unitizerTests-class}`} object
or a straight up expression, though in most cases it should be the
latter.
}
\note{
you can only do this once for a `unitizer`.
}
\keyword{internal}

