% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/test.R
\docType{methods}
\name{+,unitizerTests,unitizerSectionExpressionOrExpression-method}
\alias{+,unitizerTests,unitizerSectionExpressionOrExpression-method}
\title{Add More Tests to \code{`\link{unitizerTests-class}`}}
\usage{
\S4method{+}{unitizerTests,unitizerSectionExpressionOrExpression}(e1, e2)
}
\value{
unitizerTests-class
}
\description{
Will overwrite the call at the current index with the contents of the
expression passed as the \code{`e2`} argument, unless current index is 0L
in which case will just append the expressions.  The intended use for this
is when the evaluation of one of the tests produces a
\code{`\link{unitizerSectionExpression-class}`} that should then replace the
call.
}
\details{
The index of the \code{`\link{unitizerTests-class}`} object will be set so that
the next call to \code{`\link{nextItem,unitizerList-method}`} will return the first
call that was added.
}
\keyword{internal}

