% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.R
\name{pre_load}
\alias{pre_load}
\title{Preload Objects Generated By Files Into An Environment}
\usage{
pre_load(source, env.par = .GlobalEnv)
}
\arguments{
\item{source}{the directory of files to load, or a list to turn into
an environment}

\item{env.par}{the parent of the environment to source the files into}
}
\value{
an environment containing the objects created by the sourcing of the
  files in \code{source}
}
\description{
Preload Objects Generated By Files Into An Environment
}
\keyword{internal}

