% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.bullet}
\alias{print.bullet}
\title{Print Methods for \code{`\link{UL}`} and \code{`\link{OL}`} objects}
\usage{
\method{print}{bullet}(x, width = 0L, ...)
}
\arguments{
\item{x}{object to print}

\item{width}{integer how many characters to wrap at, if set to 0 will auto
detect width with \code{getOptions("width")}}
}
\value{
invisibly a character vector with one element per line printed
}
\description{
Print Methods for \code{`\link{UL}`} and \code{`\link{OL}`} objects
}
\keyword{internal}

