% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.H1}
\alias{print.H1}
\alias{print.H2,}
\alias{print.H3,}
\alias{print.header}
\title{Print a header}
\usage{
\method{print}{H1}(x, ...)
}
\arguments{
\item{x}{a 1 length character vector}

\item{margin}{one of "both", "top", "bottom", "none", weather to add newlines at top or bottom}
}
\value{
1 length character vector
}
\description{
Print a header
}
\keyword{internal}

