% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/asciiml.R
\name{print.header}
\alias{print.H2,}
\alias{print.H3,}
\alias{print.header}
\title{Print a header}
\usage{
\method{print}{header}(x, margin = "bottom", ...)
}
\arguments{
\item{x}{a 1 length character vector}

\item{margin}{one of "both", "top", "bottom", "none", weather to add newlines at top or bottom}
}
\value{
1 length character vector
}
\description{
Print a header
}
\keyword{internal}

