% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/browse.struct.R
\name{processInput}
\alias{processInput}
\title{Based on User Input, Return Either Reference Or New Items}
\usage{
processInput(x, ...)
}
\description{
Translates "Y", "N", etc. into c("A", "B", "C"), where "A" means return value
from new item list, "B" return value from old item list (the original store)
and "C" means return NULL.
}
\keyword{internal}

