% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/text.R
\name{Rdiff_obj}
\alias{Rdiff_obj}
\title{Do a \code{\link{tools::Rdiff}} Between R Objects}
\usage{
Rdiff_obj(from, to, ...)
}
\arguments{
\item{from}{an R object}

\item{to}{another R object}

\item{...}{passed on to \code{Rdiff}}
}
\value{
whatever \code{Rdiff} returns
}
\description{
Just a wrapper that saves the \code{print} / \code{show} representation of an
object to a temp file and then runs \code{\link{tools::Rdiff}} on them.
}
\seealso{
\code{\link{tools::Rdiff}}
}

