% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/search.R
\name{reattach}
\alias{move_on_path}
\alias{reattach}
\title{Path manipulation functions}
\usage{
reattach(pos, name, type, data, extra = NULL)

move_on_path(new.pos, old.pos)
}
\description{
Reattaches a previously detached package to the search path
}
\keyword{internal}

