% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{refSections,unitizer,unitizer-method}
\alias{refSections,unitizer,unitizer-method}
\title{Extract Reference Section Data}
\usage{
\S4method{refSections}{unitizer,unitizer}(x, y)
}
\arguments{
\item{x}{the new unitizer that will be stored with the reference tests}

\item{y}{the unitizer that will be used to generate the sections}
}
\description{
Using one unitizer with existing new items, and another unitizer that we
just created from it by pulling out the tests we intend to keep, recreate
the sections for the tests we intend to keep.
}
\details{
This isn't super robust as we're not ensuring that the two unitizers used
here are related in any way.  Would be better to have something that does
this properly...
}
\keyword{internal}

