\docType{methods}
\name{registerItem,testor,testorItem-method}
\alias{registerItem,testor,testorItem-method}
\alias{testItem,testor,testorItem-method}
\title{Helper Methods for Adding Items to \code{`\link{testor-class}`} Object}
\usage{
\S4method{registerItem}{testor,testorItem}(e1, e2, ...)
}
\description{
Helper Methods for Adding Items to
\code{`\link{testor-class}`} Object
}
\seealso{
\code{`\link{+,testor,testorItem-method}`}
}

