% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{relativize_path}
\alias{relativize_path}
\title{Simplify a Path As Much as Possible to Working Directory}
\usage{
relativize_path(path, wd = NULL, only.if.shorter = TRUE)
}
\arguments{
\item{wd}{NULL or character(1L) resolving to a directory, if NULL will be
resolved to either \code{getwd} or what \code{getwd} was at the beginning
of a \code{unitizer} run}

\item{only.if.shorter}{logical(1L) whether to relativize only if the
resulting \code{path} is shorter than the input}
}
\description{
Simplify a Path As Much as Possible to Working Directory
}
\keyword{internal}

