% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/upgrade.R
\name{renameSlot}
\alias{renameSlot}
\title{Rename a slot}
\usage{
renameSlot(object, old.name, new.name)
}
\description{
Basically assumes old name exists in object, but not in new class definition
}
\keyword{internal}

