% Generated by roxygen2 (4.0.2): do not edit by hand
\name{repair_environments}
\alias{repair_environments}
\title{Repair Environment Chains}
\usage{
repair_environments(x)
}
\arguments{
\item{x}{either a unitizer, or a store id (see \code{`\link{unitize}`})}
}
\value{
a \code{`unitizer`} object
}
\description{
In theory should never be needed, but use in case you get errors about
corrupted environments.  You should only use this if you get an error
telling you to use it.
}
\seealso{
\code{`\link{unitize}`}
}

