% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unitize.core.R
\name{reset_and_unshim}
\alias{reset_and_unshim}
\title{Helper function for global state stuff}
\usage{
reset_and_unshim(global)
}
\description{
Maybe this should be a global method?
}
\keyword{internal}

