% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zeroenv.R
\name{reset_packenv}
\alias{reset_packenv}
\title{Clears out Global State}
\usage{
reset_packenv()
}
\value{
the package environment we use as a global variable
}
\description{
Should be called by `unitize` each time
}
\keyword{internal}

