% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/browse.R
\docType{methods}
\name{reviewNext,unitizerBrowse-method}
\alias{reviewNext,unitizerBrowse-method}
\title{Bring up Review of Next test}
\usage{
\S4method{reviewNext}{unitizerBrowse}(x, ...)
}
\description{
Generally we will go from one test to the next, where the next test is
determined by the value of \code{x@last.id}.  This means it is possible
to affect the browsing order by modifying \code{x@last.id}.
}
\details{
This method is in charge of displaying all the output for review.
}
\keyword{internal}

