% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/text.R
\name{screen_out}
\alias{screen_out}
\title{Display helper function}
\usage{
screen_out(txt, max.len = getOption("unitizer.test.out.lines"),
  file = stdout())
}
\description{
Display helper function
}
\keyword{internal}

