% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/global.R
\name{search_as_envs}
\alias{search_as_envs}
\title{Get Current Search Path as List of Environments}
\usage{
search_as_envs()
}
\description{
Internal utility function.  Loaded namespaces attached as an attribute.
Probably should be an S4 class.
}
\details{
This has to be in this file, and not in R/search.R for teh setClass for the
state funs object.
}
\keyword{internal}

