% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/search.R
\name{search_path_attrs}
\alias{search_path_attrs}
\title{Generate An Identifier out of SP objects}
\usage{
search_path_attrs(x = NULL)
}
\description{
Not perfect, by any means, but necessary because we can't directly compare
the search path environments as they change on detach / re-attach
}
\keyword{internal}

