% Generated by roxygen2 (4.0.2): do not edit by hand
\name{search_path_restore}
\alias{search_path_restore}
\title{Restore Search Path to State Before \code{`search_path_trim`}}
\usage{
search_path_restore()
}
\value{
TRUE on success, FALSE on failure, invisibly
}
\description{
Undoes \code{`search_path_trim`}
}
\seealso{
\code{`\link{search_path_trim}`}  \code{`\link{search}`}
}
\keyword{internal}

