% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zeroenv.R
\name{search_path_trim}
\alias{search_path_trim}
\title{Restore Search Path to Bare Bones R Default}
\usage{
search_path_trim(keep = c("package:unitizer", "tools:rstudio"))
}
\arguments{
\item{keep}{character names of packages/objects to keep in search path;
note that base packages (see .unitizer.base.packages) that come typically
pre attached are always kept.  The \code{`keep`} packages are an addition
to those.}
}
\value{
invisibly TRUE on success, FALSE on failure
}
\description{
\code{`search_path_trimp`} attempts to recreate a clean environment by
unloading all packages and objects that are not loaded by default in the
default R  configuration.
}
\details{
Note this does not unload namespaces, but rather just detaches them from
the namespace

\code{`tools:rstudio`} is kept in search path as the default argument because
it isn't possible to cleanly unload and reload it because \code{`attach`}
actually attaches a copy of it's argument, not the actual object, and that
causes problems for that search path item.
}
\seealso{
\code{`\link{search_path_restore}`}  \code{`\link{search}`}
}
\keyword{internal}

