% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/search.R
\name{search_path_unique_id}
\alias{search_path_unique_id}
\title{Make Unique IDs For Search Path Object}
\usage{
search_path_unique_id(search.path.objs)
}
\description{
adds ".1", ".2" etc if there are non-unique values, but first occurence is
not modified so we can match between a list with duplicates and one without.
}
\keyword{internal}

