% Generated by roxygen2 (4.0.2): do not edit by hand
\name{set_trace}
\alias{set_trace}
\title{Recompute a Traceback}
\usage{
set_trace(trace)
}
\arguments{
\item{trace}{a list of type generated by sys.calls()}
}
\value{
TRUE (only purpose of this is side effect)
}
\description{
Used for cases where the trace isn't generated because the error was run within
a handling loop, but we still want the trace so we can emulate command line
behavior.
}
\details{
This will modify the .Traceback system variable (see \code{`\link{traceback}`}
documentation).

Assumption right now is that the outer most call to \code{`withCallingHandlers`}
is the baseline level from which we want to repor the traceback.
}
\keyword{internal}

