% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/options.R
\name{setOptIfNotSet}
\alias{setOptIfNotSet}
\title{Set Option if Not Set}
\usage{
setOptIfNotSet(x, value)
}
\arguments{
\item{x}{the name of the option to set}

\item{value}{the value to set the option to}
}
\value{
the result of calling \code{option} if option is not set, or NULL
  invisibly if it is already set
}
\description{
Utility function used by .onLoad to set a global option if it is not set
already.  This is equivalent to conditionally using \code{option} to set
options that are not defined yet.
}

