% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conditions.R
\docType{methods}
\name{show,conditionList-method}
\alias{show,conditionList-method}
\title{Prints A list of Conditions}
\usage{
\S4method{show}{conditionList}(object)
}
\arguments{
\item{object}{a \code{`conditionList`} object (list of conditions)}
}
\value{
object, invisibly
}
\description{
Prints A list of Conditions
}

