% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{show,unitizerBrowse-method}
\alias{show,unitizerBrowse-method}
\title{Display Summary of Tests and User Decisions}
\usage{
\S4method{show}{unitizerBrowse}(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
Used to help navigate tests.  Will only show reviewed tests because
implementing the ability to skip ahead has several annoying implications
that we did not want to support (need to check that eventually all tests
are reviewed, etc.)
}
\keyword{internal}

