% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{show,unitizerChanges-method}
\alias{show,unitizerChanges-method}
\title{Print Out A Summary Of the Changes}
\usage{
\S4method{show}{unitizerChanges}(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
Print Out A Summary Of the Changes
}
\keyword{internal}

