% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show,unitizerItem-method}
\alias{show,unitizerItem-method}
\title{Display a \code{`\link{unitizerItem-class}`} Object}
\usage{
\S4method{show}{unitizerItem}(object)
}
\description{
Highly summarized view of the unitizer object.
}
\keyword{internal}

