% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.sub.R
\docType{methods}
\name{show,unitizerItemTestsErrorsDiff-method}
\alias{show,unitizerItemTestsErrorsDiff-method}
\title{Show Method for unitizerItemTestsErrorsDiff objects}
\usage{
\S4method{show}{unitizerItemTestsErrorsDiff}(object)
}
\description{
Show Method for unitizerItemTestsErrorsDiff objects
}
\keyword{internal}
