% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.sub.R
\name{show,unitizerItemTestsErrorsDiffs-method}
\alias{show,unitizerItemTestsErrorsDiffs-method}
\title{Show Method for unitizerItemTestsErrorsDiffs objects}
\usage{
\S4method{show}{unitizerItemTestsErrorsDiffs}(object)
}
\description{
Show Method for unitizerItemTestsErrorsDiffs objects
}
\keyword{internal}
