% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unitizer.R
\docType{methods}
\name{show,unitizerObjectListSummary-method}
\alias{show,unitizerObjectListSummary-method}
\title{Display method}
\usage{
\S4method{show}{unitizerObjectListSummary}(object)
}
\description{
Display method
}
\keyword{internal}

