% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show,unitizerSummary-method}
\alias{show,unitizerSummary-method}
\title{Display Unitizer Summary}
\usage{
\S4method{show}{unitizerSummary}(object)
}
\arguments{
\item{object}{the object to show}
}
\description{
Display Unitizer Summary
}
\keyword{internal}

