% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\docType{methods}
\name{show.conditionList}
\alias{show.conditionList}
\title{Prints A list of Conditions}
\usage{
\S4method{show}{conditionList}(object)
}
\arguments{
\item{object}{a \code{\link{conditionList}} object (list of conditions)}
}
\value{
object, invisibly
}
\description{
S4 method for \code{\link{conditionList}} objects.
}
\seealso{
\code{\link{conditionList}}
}
