% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{sizeUntz}
\alias{sizeUntz}
\title{Utility To Examine Object Size}
\usage{
sizeUntz(x, ...)
}
\description{
Funny name is just to avoid conflicts with functions with same names in
other packages.
}
\keyword{internal}
