% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.R
\name{source_files}
\alias{source_files}
\title{Run \code{sys.source} On All Provided Files}
\usage{
source_files(files, env.par, pattern = "\\\\.[rR]$")
}
\arguments{
\item{files}{character() pointing to files or directories}

\item{env}{an environment}
}
\value{
environment, or a character message explaining why it failed
}
\description{
Sorts them before sourcing on the paths as specified in \code{files}, returns
an environment containing any objects created by the code in the source
files.  If a file is a directory, the files in that directory are sourced,
though only top level files in that directory are sourced.
}
\keyword{internal}

