% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.R
\name{source_many}
\alias{source_many}
\title{Run \code{sys.source} on a list of files}
\usage{
source_many(files, env, allow.side.effects = FALSE)
}
\arguments{
\item{files}{character()}

\item{env}{an environment}

\item{allow.side.effects}{checks for changes to \code{options} or working
directory, and fails if there are any (though note that original values
are not restored); should probably add this check to main \code{unitizer}?}
}
\value{
NULL, parameter \code{env} is modified by reference (note this
  happens even if we hit an error before getting through all the files)
}
\description{
Sorts them before sourcing on the paths as specified in \code{files}
}
\keyword{internal}

