% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{str_reduce_unique}
\alias{str_reduce_unique}
\title{Remove Common Characters From Values in a Vector}
\usage{
str_reduce_unique(x, from = "left")
}
\arguments{
\item{x}{character the vector to make more unique}

\item{from}{the direction to remove common elements from}
}
\description{
Note that one length \code{x} is a degenerate case that returns "".
}
\keyword{internal}

