% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{[[,unitizerList,subIndex-method}
\alias{[[,unitizerList,subIndex-method}
\title{Subsetting Method for \code{`\link{unitizerList-class}`}}
\usage{
\S4method{[[}{unitizerList,subIndex}(x, i)
}
\description{
Subsetting Method for \code{`\link{unitizerList-class}`}
}
\keyword{internal}

