% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/browse.struct.R
\docType{methods}
\name{[,unitizerBrowse,subIndex,missing,missing-method}
\alias{[,unitizerBrowse,subIndex,missing,missing-method}
\title{Subset A \code{unitizerBrowse} Object}
\usage{
\S4method{[}{unitizerBrowse,subIndex,missing,missing}(x, i)
}
\description{
Used primarily to confirm actions on multiple items.  Note this means ids
are no longer continuous, something that we assume when we cycle through
items.  Need to think about this a bit...
}
\details{
Generally, be careful about using a subsetted browse object as you would
a non-subsetted one until we get around to making cycling more robust.

Finally, note that this conflicts with the underlying nature of a
\code{unitizerList} since we're overriding the \code{[} method.  All of this
is caused by the nested nature of sections and sub-sections, which is
feeling like a worse design decision every time I look at it.  Note also that
something like \code{ubobj[4]} and \code{ubobj[[4]]} will likely return
completely different things as in the former we are subsetting based on the
order implied by \code{ubobj@mapping}, whereas in the latter we're directly
pulling out an entire section.  Obviously not ideal, but since this is
internal we're going to ignore the problem for now.
}
\keyword{internal}

