% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/browse.struct.R
\docType{methods}
\name{[,unitizerBrowseSubSection,subIndex,missing,missing-method}
\alias{[,unitizerBrowseSubSection,subIndex,missing,missing-method}
\title{Subset a \code{unitizerBrowseSubSection} Object}
\usage{
\S4method{[}{unitizerBrowseSubSection,subIndex,missing,missing}(x, i)
}
\description{
Subset a \code{unitizerBrowseSubSection} Object
}
\keyword{internal}

