% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/list.R
\docType{methods}
\name{[<-,unitizerList,subIndex-method}
\alias{[<-,unitizerList,subIndex-method}
\title{Replace Method for \code{`\link{unitizerList-class}`}}
\usage{
\S4method{[}{unitizerList,subIndex}(x, i) <- value
}
\description{
Replace Method for \code{`\link{unitizerList-class}`}
}
\keyword{internal}

