% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/text.R
\name{substr_cons}
\alias{substr_cons}
\title{Substring To a Length, but end In Consonant}
\usage{
substr_cons(x, stop, justify = "left")
}
\arguments{
\item{x}{character vector to substring}

\item{stop}{integer max number of characters}

\item{justify}{character(1L) passed on to format}
}
\description{
Substring To a Length, but end In Consonant
}
\keyword{internal}

