% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{summ_matrix_to_text}
\alias{summ_matrix_to_text}
\title{Convert A Matrix of Test Outcomes for Display}
\usage{
summ_matrix_to_text(mx, from = "right", width = getOption("width"),
  show.nums = TRUE)
}
\description{
Used by \code{show} methods for both \code{unitizerSummary} and
\code{unitizerSummaryList}
}
\keyword{internal}

