% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{summary,unitizer-method}
\alias{summary,unitizer-method}
\title{Summarize Results}
\usage{
\S4method{summary}{unitizer}(object, ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Summarize Results
}
\keyword{internal}

