% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.sub.R
\docType{methods}
\name{summary,unitizerItemTestsErrors-method}
\alias{summary,unitizerItemTestsErrors-method}
\title{Summary Method for unitizerItemTestsErrors Objects}
\usage{
\S4method{summary}{unitizerItemTestsErrors}(object, ...)
}
\arguments{
\item{object}{a \code{unitizerItemTestsErrors} object}
}
\value{
NULL, invisibly
}
\description{
Summary Method for unitizerItemTestsErrors Objects
}

