% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unitizer.R
\docType{methods}
\name{summary,unitizerObjectList-method}
\alias{summary,unitizerObjectList-method}
\title{Summary method}
\usage{
\S4method{summary}{unitizerObjectList}(object, silent = FALSE, ...)
}
\description{
Summary method
}
\keyword{internal}

