% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{testor}
\alias{testor}
\alias{testor-package}
\title{testor}
\description{
Stored unit tests.  Running tests compares expressions against previously
run versions that were stored as references.  This allows detection of any
changes in behavior in the code as in theory the new evaluation should
match the reference stored evaluation.
}

