% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{testorChanges-class}
\alias{testorChanges-class}
\title{Summary of Changes}
\description{
Changes arise any time a user, through the interactive testor mode,
requests the storing of a change (accept new version of failed test,
add new test, remove old test).
}

