% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{testorItemTestError-class}
\alias{testorItemTestError-class}
\alias{testorItemTestsErrors-class}
\alias{testorItemsTestsErrors-class,}
\title{Stores Errors from Evaluating New vs. Ref Comparisons}
\description{
There various nested objects involved:
\itemize{
  \item \code{`testorItemTestError`} contains the error produced from a comparison
  \item \code{`testorItemTestsErrors`} aggregates the errors for each slot of an item
  \item \code{`testorItemsTestsErrors`} aggregates all the errors for the
     \code{`\link{testor-class}`} object
}
}
\keyword{internal}

