% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{testorItemTestFun-class}
\alias{testorItemTestFun-class}
\title{Validates Functions for Use in New vs. Reference Object Comparison}
\description{
Validates Functions for Use in New vs. Reference Object Comparison
}
\keyword{internal}

