% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{testorItemTests-class}
\alias{testorItemTests-class}
\title{Virtual Class To Enforce Slots on Subclasses}
\description{
Used for \code{`\link{testorItemsTestFuns}`} and \code{`\link{testorItemTestsErrors}`}
}
\keyword{internal}

