% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/translate.R
\name{testthat_match_call}
\alias{testthat_match_call}
\title{Pull out parameter from call}
\usage{
testthat_match_call(call, fun, target.params)
}
\arguments{
\item{target.params}{parameters required in the matched call}
}
\description{
Pull out parameter from call
}
\keyword{internal}

