% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/translate.R
\name{testthat_translatable_fun}
\alias{testthat_translatable_fun}
\title{Determine if a Function Meets Translatable Profile}
\usage{
testthat_translatable_fun(x)
}
\description{
Determine if a Function Meets Translatable Profile
}
\keyword{internal}

