% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/browse.R
\name{toggleReeval}
\alias{toggleReeval}
\title{Re-eval toggling, only b/c we need to do it in a couple of places'}
\usage{
toggleReeval(x, ...)
}
\description{
Re-eval toggling, only b/c we need to do it in a couple of places'
}
\keyword{internal}

