% Generated by roxygen2 (4.0.1): do not edit by hand
\name{UL}
\alias{OL}
\alias{UL}
\title{Create List Objects}
\usage{
UL(x)
}
\arguments{
\item{x}{character vector of items to make a list out of}
}
\value{
OL/UL object
}
\description{
Turns a character vector into list items.
}
\details{
Currently doesn't support nested lists, but this might be added in the future.
}
\keyword{internal}

