% Generated by roxygen2 (4.0.2): do not edit by hand
\name{uncomment}
\alias{uncomment}
\title{Remove any comment attributes}
\usage{
uncomment(lang)
}
\description{
Used by the internal deparse functions.  Really removes all attributes.
Resorting to desperate measures due to the reference like behavior of
expressions and messing with their attributes, most likely due to the
srcref style environment attributes.
}
\keyword{internal}

