% Generated by roxygen2 (4.0.2): do not edit by hand
\name{unitize}
\alias{unitize}
\title{Unitize an R Test Script}
\usage{
unitize(test.file, store.id = sub("\\\\.[Rr]$", ".unitizer", test.file),
  interactive.mode = interactive(), env.clean = TRUE,
  search.path.clean = getOption("unitizer.search.path.clean"),
  search.path.keep = c("tools:rstudio", "package:unitizer"))
}
\arguments{
\item{test.file}{path to the file containing tests}

\item{store.id}{a folder to store the \code{`unitizer`} objects in; will auto-
generate to a folder at the same location as the test file with the same
name as the testfile, except ending in \code{`.unitizer`} instead of \code{`.R`}}

\item{interactive.mode}{logical(1L) whether to run in interactive mode}

\item{env.clean}{TRUE or environment, if TRUE tests are run in a clean
environment, if an environment they are run with that environment as the
parent.}

\item{search.path.clean}{logical(1L) if TRUE all items on the search path that
are not part of a clean R session are detached prior to running tests.  Note
namespaces for detached packages remain loaded.  Additionally, the search
path is restored to its initial state upon exiting \code{`unitizer`} so any
packages added/removed, or objects attached/detached from search path are
restored to original state.  This feature is somewhat experimental and is
disabled by default, though this will likely change in the future.  See
"Reproducible Tests" vignette for details.}

\item{search.path.keep}{character any additional items on the search path
to keep attached; has no effect unless \code{`search.path.clean`} is TRUE}
}
\description{
Turns a standard R script into unit tests by evaluating the expressions and
storing them along with their resuls.  Re-running \code{`unitize`} then
checks that the values remain unchanged.  See "unitizer" vignette for more
details.
}
\details{
You can run \code{`unitize`} from the command line, or you can place one or
more \code{`unitize`} calls in an R file and source that.
}
\seealso{
\code{`\link{get_store}`}
}

