% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unitize.core.R
\name{unitize_browse}
\alias{unitize_browse}
\title{Run User Interaction And \code{unitizer} Storage}
\usage{
unitize_browse(unitizers, mode, interactive.mode, force.update, auto.accept)
}
\arguments{
\item{unitizers}{list of \code{unitizer} objects}

\item{mode}{character(1L) one of "review" or "unitize"}

\item{interactive.mode}{logical(1L) whether to run in interactive mode (
request user input when needed) or not (error if user input is required,
e.g. if all tests do not pass).}

\item{force.update}{whether to store unitizer}

\item{auto.accept}{character(XL) ADVANCED USE ONLY: YOU CAN EASILY DESTROY
YOUR \code{unitizer} WITH THIS; whether to auto-accept tests without
prompting, use values in \code{c("new", "failed", "deleted", "error")} to
specify which type(s) of test you wish to auto accept (i.e. same as typing
\code{"Y"} at the \code{unitizer} prompt) or empty character vector to turn
off (default)}
}
\description{
Run User Interaction And \code{unitizer} Storage
}
\keyword{internal}

