% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unitize.core.R
\name{unitize_eval}
\alias{unitize_eval}
\title{Evaluate User Tests}
\usage{
unitize_eval(tests.parsed, unitizers)
}
\arguments{
\item{tests.parsed}{a list of expressions}

\item{unitizers}{a list of \code{unitizer} objects of same length as \code{tests.parsed}}

\item{which}{integer which of \code{unitizer}s to actually eval, all get
summary status displayed to screen}
}
\value{
a list of unitizers
}
\description{
Evaluate User Tests
}
\keyword{internal}

