% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{.unitizer.opts.base}
\alias{.unitizer.opts.base}
\alias{.unitizer.opts.sys}
\title{Variables Controlling Global Option State During \code{unitizer}}
\format{\preformatted{List of 45
 $ add.smooth           : logi TRUE
 $ browserNLdisabled    : logi FALSE
 $ CBoundsCheck         : logi FALSE
 $ check.bounds         : logi FALSE
 $ citation.bibtex.max  : num 1
 $ continue             : chr "+ "
 $ contrasts            : Named chr [1:2] "contr.treatment" "contr.poly"
  ..- attr(*, "names")= chr [1:2] "unordered" "ordered"
 $ defaultPackages      : chr [1:6] "datasets" "utils" "grDevices" "graphics" ...
 $ demo.ask             : chr "default"
 $ deparse.cutoff       : int 60
 $ device.ask.default   : logi FALSE
 $ digits               : int 7
 $ echo                 : logi TRUE
 $ encoding             : chr "native.enc"
 $ NA                   : NULL
 $ example.ask          : chr "default"
 $ expressions          : int 5000
 $ help.search.types    : chr [1:3] "vignette" "demo" "help"
 $ help.try.all.packages: logi FALSE
 $ internet.info        : num 2
 $ keep.source          : logi TRUE
 $ keep.source.pkgs     : logi FALSE
 $ locatorBell          : logi TRUE
 $ max.print            : int 99999
 $ na.action            : chr "na.omit"
 $ nwarnings            : int 50
 $ OutDec               : chr "."
 $ prompt               : chr "> "
 $ repos                : Named chr "http://cran.r-project.org"
  ..- attr(*, "names")= chr "CRAN"
 $ rl_word_breaks       : chr " \\t\\n\\"\\\\'`><=\%;,|&\{()\}"
 $ scipen               : num 0
 $ show.coef.Pvalues    : logi TRUE
 $ show.error.messages  : logi TRUE
 $ show.signif.stars    : logi TRUE
 $ str                  :List of 3
  ..$ strict.width: chr "no"
  ..$ digits.d    : num 3
  ..$ vec.len     : num 4
 $ str.dendrogram.last  : chr "`"
 $ stringsAsFactors     : logi TRUE
 $ timeout              : num 60
 $ ts.eps               : num 1e-05
 $ ts.S.compat          : logi FALSE
 $ useFancyQuotes       : logi TRUE
 $ verbose              : logi FALSE
 $ warn                 : num 0
 $ warning.length       : int 1000
 $ width                : int 80
}}
\usage{
.unitizer.opts.base

.unitizer.opts.sys
}
\description{
These variables are used when setting options state to the highest
reproducibility setting to minimize odds that test failures are caused
by changes to global options.
}
\details{
If running in highest reproducibily mode, all options except those that match
those set in \code{getOption("unitizer.opts.asis")} are set to NULL.  Then,
we use \code{getOption("unitizer.opts.base")} to set base line options that
are typically loaded in a fresh vanilla R session.

Any packages loaded during tests should set their own options.
\code{unitizer} unloads packages so that each new invocation of the package
runs the \code{onload} scripts and resets options.  Because of this, any
packages that you add to \code{getOption("unitizer.search.path.keep")} must
have their options included in \code{getOption("unitizer.opts.asis")} as
otherwise those options would be unset, but not reset since \code{unitizer}
will not unload and reload the package.

Occasionally you will run into packages that cannot be unloaded.  If your
tests involve such a package you can either add it to
\code{getOption("unitizer.search.path.keep")} and its options to
\code{getOption("unitizer.opts.asis")}, or turn off the reproducibility
setting for options.
}
\keyword{datasets}

