% Generated by roxygen2 (4.0.2): do not edit by hand
\name{unitize_core}
\alias{unitize_core}
\title{Runs The Basic Stuff}
\usage{
unitize_core(test.file, store.id, interactive.mode, env.clean,
  search.path.clean, search.path.keep, force.update)
}
\description{
Used by both \code{`\link{unitize}`} and \code{`\link{review,unitizer-method}`}
to launch the interactive interface for reviewing tests.
}
\details{
Right now we distinguish in what mode we're running based on whether
\code{`test.file`} is NULL (review mode) vs. not (unitize mode), which isn't
very elegant, but whatevs.  This has implications for the parsing / evaluation
step, as well as how the \code{`unitizerBrowse`} object is constructed.  Otherwise
stuff is mostly the same.

Cleary there is a trade-off in increased code complexity to handle both types
of code, vs duplication.  Not ideal, but tasks are so closely related and
there is so much common overhead, that the central function makes sense.
Also, since unfortunately we're relying on side-effects for some features, and
`on.exit` call for safe operation, it is difficult to truly modularize.
}
\keyword{internal}

