% Generated by roxygen2 (4.0.1): do not edit by hand
\name{unitizer_ls}
\alias{unitizer_ls}
\title{An `ls` Like Function}
\usage{
unitizer_ls(name, pos = -1L, envir = as.environment(pos),
  all.names = FALSE, pattern)
}
\value{
list of object names, or a list with environments containing the objects
}
\description{
Much like `ls`, except that it is designed to crawl up the \code{`.new`} and
\code{`.ref`} environments and display all the objects.
}
\details{
This is used in \code{`\link{browse,unitizer-method}`}, and is re-assigned to
\code{`ls`} for use in the \code{`unitizer`} prompt environment.
}
\keyword{internal}

