% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/class_unions.R, R/list.R,
%   R/item.R, R/item.sub.R, R/test.R, R/unitizer.R, R/browse.struct.R, R/load.R,
%   R/state.R, R/state.compare.R, R/unitizer.add.R
\docType{methods}
\name{show,unitizerDummy-method}
\alias{show,unitizerDummy-method}
\alias{as.integer,unitizerGlobalIndices-method}
\alias{unitizer_s4method_doc}
\alias{length,unitizerList-method}
\alias{[,unitizerList,subIndex,missing,missing-method}
\alias{[[,unitizerList,subIndex-method}
\alias{[<-,unitizerList,subIndex-method}
\alias{[[<-,unitizerList,subIndex-method}
\alias{as.list,unitizerList-method}
\alias{append,unitizerList,ANY-method}
\alias{c,unitizerList-method}
\alias{append,factor,factor-method}
\alias{names,unitizerList-method}
\alias{names<-,unitizerList-method}
\alias{initialize,unitizerItem-method}
\alias{show,unitizerItem-method}
\alias{+,unitizerItems,unitizerItemOrNULL-method}
\alias{+,unitizerItems,unitizerItems-method}
\alias{show,unitizerItemTestsErrors-method}
\alias{+,unitizerTests,unitizerSectionExpressionOrExpression-method}
\alias{show,unitizerSummary-method}
\alias{initialize,unitizer-method}
\alias{length,unitizer-method}
\alias{summary,unitizer-method}
\alias{summary,unitizerObjectList-method}
\alias{show,unitizerObjectListSummary-method}
\alias{as.character,unitizer-method}
\alias{show,unitizerBrowse-method}
\alias{as.character,unitizerBrowse-method}
\alias{as.data.frame,unitizerBrowse-method}
\alias{+,unitizerBrowse,unitizerBrowseSection-method}
\alias{length,unitizerBrowseSubSection-method}
\alias{[,unitizerBrowse,subIndex,missing,missing-method}
\alias{[,unitizerBrowseSubSection,subIndex,missing,missing-method}
\alias{+,unitizerBrowseSection,unitizerBrowseSubSection-method}
\alias{show,unitizerLoadFail-method}
\alias{as.character,unitizerInPkg-method}
\alias{show,unitizerInPkg-method}
\alias{show,unitizerState-method}
\alias{all.equal,unitizerDummy,unitizerDummy-method}
\alias{all.equal,unitizerDummy,ANY-method}
\alias{all.equal,ANY,unitizerDummy-method}
\alias{all.equal,unitizerStateRaw,unitizerStateRaw-method}
\alias{+,unitizer,unitizerSection-method}
\alias{+,unitizer,unitizerTestsOrExpression-method}
\alias{+,unitizer,unitizerItems-method}
\alias{+,unitizer,unitizerItem-method}
\alias{+,unitizer,unitizerItemTestsErrors-method}
\title{Documentation Block for Internal S4 Methods}
\usage{
\S4method{show}{unitizerDummy}(object)

\S4method{as.integer}{unitizerGlobalIndices}(x, ...)

\S4method{length}{unitizerList}(x)

\S4method{[}{unitizerList,subIndex,missing,missing}(x, i)

\S4method{[[}{unitizerList,subIndex}(x, i)

\S4method{[}{unitizerList,subIndex}(x, i) <- value

\S4method{[[}{unitizerList,subIndex}(x, i) <- value

\S4method{as.list}{unitizerList}(x, ...)

\S4method{append}{unitizerList,ANY}(x, values, after = length(x))

\S4method{c}{unitizerList}(x, ..., recursive = FALSE)

\S4method{append}{factor,factor}(x, values, after = length(x))

\S4method{names}{unitizerList}(x)

\S4method{names}{unitizerList}(x) <- value

\S4method{initialize}{unitizerItem}(.Object, ...)

\S4method{show}{unitizerItem}(object)

\S4method{+}{unitizerItems,unitizerItemOrNULL}(e1, e2)

\S4method{+}{unitizerItems,unitizerItems}(e1, e2)

\S4method{show}{unitizerItemTestsErrors}(object)

\S4method{+}{unitizerTests,unitizerSectionExpressionOrExpression}(e1, e2)

\S4method{show}{unitizerSummary}(object)

\S4method{initialize}{unitizer}(.Object, ...)

\S4method{length}{unitizer}(x)

\S4method{summary}{unitizer}(object, silent = FALSE, ...)

\S4method{summary}{unitizerObjectList}(object, silent = FALSE, ...)

\S4method{show}{unitizerObjectListSummary}(object)

\S4method{as.character}{unitizer}(x, ...)

\S4method{show}{unitizerBrowse}(object)

\S4method{as.character}{unitizerBrowse}(x, width = 0L, ...)

\S4method{as.data.frame}{unitizerBrowse}(x, row.names = NULL,
  optional = FALSE, ...)

\S4method{+}{unitizerBrowse,unitizerBrowseSection}(e1, e2)

\S4method{length}{unitizerBrowseSubSection}(x)

\S4method{[}{unitizerBrowse,subIndex,missing,missing}(x, i)

\S4method{[}{unitizerBrowseSubSection,subIndex,missing,missing}(x, i)

\S4method{+}{unitizerBrowseSection,unitizerBrowseSubSection}(e1, e2)

\S4method{show}{unitizerLoadFail}(object)

\S4method{as.character}{unitizerInPkg}(x, ...)

\S4method{show}{unitizerInPkg}(object)

\S4method{show}{unitizerState}(object)

\S4method{all.equal}{unitizerDummy,unitizerDummy}(target, current, ...)

\S4method{all.equal}{unitizerDummy,ANY}(target, current, ...)

\S4method{all.equal}{ANY,unitizerDummy}(target, current, ...)

\S4method{all.equal}{unitizerStateRaw,unitizerStateRaw}(target, current,
  ...)

\S4method{+}{unitizer,unitizerSection}(e1, e2)

\S4method{+}{unitizer,unitizerTestsOrExpression}(e1, e2)

\S4method{+}{unitizer,unitizerItems}(e1, e2)

\S4method{+}{unitizer,unitizerItem}(e1, e2)

\S4method{+}{unitizer,unitizerItemTestsErrors}(e1, e2)
}
\description{
R insists these need to be documented as user facing, but they are not really
so were throwing them all in here.  Actual docs are in non roxygen comments
by fun definitions.
}
\details{
Put in this file because this file is included by almost every other file
}
\keyword{internal}
