% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/browse.struct.R
\docType{class}
\name{unitizerBrowse-class}
\alias{unitizerBrowse-class}
\title{Helper Object for Browsing}
\description{
Key element here is the \code{`@mapping`} slot which is generated by
\code{`\link{+,unitizerBrowse,unitizerBrowseSection-method}`}, which allows us
to navigate all the tests.
}
\keyword{internal}

