% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{unitizerBrowseMapping-class}
\alias{unitizerBrowseMapping-class}
\title{Keeps track of All Test Review Data}
\description{
Seemed like a brilliant idea to make this an object to simplify validation,
but as result cycling through the items is incredibly annoying.  Need to
develop better ways to iterate through each item while getting all the data
here, as well as ways of easily knowing which sections/subsections are
ignored.
}
\details{
The real issue with all this stuff is that \code{`item.new`} and
\code{`item.ref`} can be NULL, and whether on or the other or neither are
NULL changes the processing logic.  Probably the thing to do is extract the
non NULL values (i.e. item.main) and store them in a list, along with a
list highlighting which of \code{`item.new`} or \code{`item.ref`} has been
picked.
}
\section{Slots}{

\describe{
\item{\code{item.id}}{unique, 1 incrementing up to total number of reviewable items}

\item{\code{item.id.rel}}{non-unique, unique within each sec/sub.sec}

\item{\code{item.id.orig}}{the original id of the item used to re-order tests in the
order they show up in the original files}

\item{\code{item.ref}}{whether a test is a reference test or not}

\item{\code{reviewed}}{whether a test has been reviewed}

\item{\code{review.val}}{what action the user decided ("N") is default}
}}
\keyword{internal}

