% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/browse.struct.R
\docType{class}
\name{unitizerBrowseResult-class}
\alias{unitizerBrowseResult-class}
\title{Return value for \code{browseUnitizerInternal}}
\description{
Return value for \code{browseUnitizerInternal}
}
\keyword{internal}

