% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/browse.struct.R
\docType{class}
\name{unitizerBrowseSection-class}
\alias{unitizerBrowseSection-class}
\title{Represents a \code{`unitizer_sect`}}
\description{
Represents a \code{`unitizer_sect`}
}
\keyword{internal}

