% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{unitizerBrowseSubSection-class}
\alias{unitizerBrowseSubSection-class}
\title{Represents A Section/Action Type when Browsing}
\description{
Represents A Section/Action Type when Browsing
}
\section{Slots}{

\describe{
\item{\code{items.new}}{the new items associated with this sub sections}

\item{\code{items.ref}}{the reference items associated with this sub sections}

\item{\code{title}}{character 1 length current test types (failed, added, removed, corrupted)}

\item{\code{prompt}}{character 1 length what to prompt the user to do}

\item{\code{actions}}{character 2 length containing c("A", "B", "C"), where "A"
means return value from new item list, "B" return value from old item
list (the original store) and "C" means return NULL.  The first value
corresponds to the action on user typing `Y`, the second the action on
user typing `N`.}

\item{\code{show.msg}}{logical whether to automatically show stderr produced during
evaluation}

\item{\code{show.out}}{logical whether to automatically show stdout produced during
evaluation}

\item{\code{show.fail}}{FALSE, or a unitizerItemsTestsErrors-class object if you want
to show the details of failure}

\item{\code{new.conditions}}{whether the items produced new conditions}
}}
\keyword{internal}

