% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/browse.struct.R
\docType{class}
\name{unitizerBrowseSubSectionFailed-class}
\alias{unitizerBrowseSubSectionFailed-class}
\title{Specific Sub-Section defaults}
\description{
Specific Sub-Section defaults
}
\keyword{internal}

