% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{unitizerItem-class}
\alias{unitizerItem-class}
\title{Full Representation Of an Evaluated \code{`unitizer`} Test}
\description{
Full Representation Of an Evaluated \code{`unitizer`} Test
}
\section{Slots}{

\describe{
\item{\code{call}}{the call that is tested}

\item{\code{reference}}{whether this is a reference or new \code{`unitizerItem`}}

\item{\code{ignore}}{whether this test should be treated as a test or just a step
towards compiling a test}

\item{\code{environment}}{the environment the test was evaluated in, should contain
all relevant objects}

\item{\code{data}}{the container for the results of the evaluation of \code{`call`}}
}}
\keyword{internal}

