% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{unitizerItemData-class}
\alias{unitizerItemData-class}
\title{Data Produced During Evaluation of \code{`unitizer`} Test}
\description{
Kept separate from the \code{`\link{unitizerItem-class}`} because these
are the slots that get compared from the new item to the reference items
which means there are a whole bunch of other classes that need to have the
same structure as this and by definining it we let those other classes
confirm they have the correct structure.
}
\keyword{internal}

