% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/item.R
\docType{class}
\name{unitizerItems-class}
\alias{unitizerItems-class}
\title{Collection of \code{`\link{unitizerItem-class}`} Objects}
\description{
Collection of \code{`\link{unitizerItem-class}`} Objects
}
\keyword{internal}

