% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/item.sub.R
\docType{class}
\name{unitizerItemTestFun-class}
\alias{unitizerItemTestFun-class}
\title{Validates Functions for Use in New vs. Reference Object Comparison}
\description{
Validates Functions for Use in New vs. Reference Object Comparison
}
\keyword{internal}

