% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/item.sub.R
\docType{class}
\name{unitizerItemTests-class}
\alias{unitizerItemTests-class}
\title{Virtual Class To Enforce Slots on Subclasses}
\description{
Virtual Class To Enforce Slots on Subclasses
}
\keyword{internal}

