% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{unitizerSection-class}
\alias{unitizerSection-class}
\alias{unitizerSectionExpression-class}
\title{Contains Representation For a Section of Tests}
\description{
\code{`unitizerSectionExpression-class`} contains the actual expressions that
belong to the section, whereas \code{`unitizerSection-class`} only contains
the meta data.  The latter objects are used within \code{`\link{unitizer-class}`},
whereas the former is really just a temporary object until we can generate
the latter.
}
\section{Slots}{

\describe{
\item{\code{title}}{1 lenght character, the name of the section}

\item{\code{details}}{character vector containing additional info on the section}

\item{\code{compare}}{functions to compare the various aspects of a \code{`\link{unitizerItem-class}`}}

\item{\code{length}}{tracks size of the section}
}}
\keyword{internal}

