% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/state.R
\docType{class}
\name{unitizerState}
\alias{unitizerState}
\alias{unitizerStateBasic-class}
\alias{unitizerStateNoOpt-class}
\alias{unitizerStateOff-class}
\alias{unitizerStatePristine-class}
\title{Interface For State Control}
\description{
You can use these classes for detailed control of how \code{unitizer} tracks
and modifies state during test evaluation and review.  These can be passed
as the \code{state} argument to \code{\link{unitize}} and
\code{\link{unitize_dir}}.
}
\details{
There are four classes defined, though their only purpose is to act as
presets since they are identical except for their prototype values:
\itemize{
  \item \code{unitizerStatePristine} is the default class and implements
    the highest level of state tracking and control
  \item \code{unitizerStateNoOpt} turns off options tracking, which makes it
    a good choice if you have a lot of packages that are added to
    \code{getOption("unitizer.namespace.keep")} so that you do not need to
    also add all their options to \code{getOption("unitizer.opts.asis")}
  \item \code{unitizerStateBasic} keeps all tracking, but at a less aggressive
    level; state is reset between each test file to the state before
    you started \code{unitize}ing so that no single test file affects
    another, but the state of your workspace, search path, etc. will affect
    all the tests
  \item \code{unitizerStateOff} state tracking is turned off
}
In addition to the preset classes, you can set any of the slots to any valid
setting (see examples).  For \code{par.env} that setting is either
\code{NULL} or an environment.  for all other slots, the settings are in
\code{0:2} and mean:
\enumerate{
  \item 0 turn off state tracking
  \item 1 track, but start with state as it was when \code{unitize} was
    called
  \item 2 track and set state to what you would typically find in a clean
    R session, with the exception of \code{random.seed}, which is
    set to \code{\link{getOption("unitizer.seed")}} (of kind "Wichmann-Hill"
    as that seed is substantially smaller than the R default).
}
}
\section{Slots}{

\describe{
\item{\code{search.path}}{one of \code{\link{0:2}}}

\item{\code{options}}{one of \code{\link{0:2}}}

\item{\code{working.directory}}{one of \code{\link{0:2}}}

\item{\code{random.seed}}{one of \code{\link{0:2}}}

\item{\code{par.env}}{\code{NULL} to use the special \code{unitizer} parent
  environment, or an environment to use as the parent environment}
}}
\note{
\code{\link{unitize_dir}} and \code{\link{unitize}} can accept
  character values instead of the classes here; these are just translated to
  the corresponding class defined here.  See the docs for the \code{state}
  parameter for those functions.
}
\examples{
\dontrun{
## use a custom environment as parent env
my.env <- new.env()
unitize(..., state=new("unitizerStatePrisitine", par.env=my.env))
## Basic, but do not track options
unitize(..., state=new("unitizerStateBasic", options=0))
}
}
\seealso{
\code{\link{unitize}}, \code{\link{unitizer.opts}}
}

