% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test.R
\docType{class}
\name{unitizerTests-class}
\alias{unitizerTests-class}
\title{Collections of Calls For Testing}
\description{
Should probably add an \code{as.unitizerTests} function...
}
\keyword{internal}

