% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search.R
\name{unitizerUniqueNames}
\alias{unitizerUniqueNames}
\title{Make Unique IDs For Search Path Object}
\usage{
unitizerUniqueNames(x, ...)
}
\description{
adds ".1", ".2" etc if there are non-unique values, but first occurence is
not modified so we can match between a list with duplicates and one without.
}
\keyword{internal}

