% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{unitzer}
\alias{unitzer}
\alias{unitzer-package}
\title{unitizer}
\description{
\code{`unitizer`} evaluates arbitrary expressions and stores them along with
the result of their evaluation.  Re-running the same `unitizer` then compares
the new results for each expression to the previously stored results.  If
there are any differences `unitizer` breaks out into an interactive mode to
allow comparison of the old evaluation to the new one, and if necessary,
update the tests.
}

