% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/search.R
\name{unload_namespaces}
\alias{unload_namespaces}
\title{Unload Namespaces}
\usage{
unload_namespaces(unload)
}
\description{
Attempts to unload namespaces in an order that avoids dependency issues
based on data from \code{getNamespaceImports}
}

