% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/upgrade.R
\docType{methods}
\name{upgrade,unitizer-method}
\alias{upgrade,unitizer-method}
\title{Alter Older unitizer Versions So They Pass Validation in Latest Version}
\usage{
\S4method{upgrade}{unitizer}(object, ...)
}
\arguments{
\item{object}{an unitizer object}

\item{...}{other arguments}
}
\value{
an upgraded unitizer object
}
\description{
Sequentially applies all applicable patches
}
\keyword{internal}

