% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/upgrade.R
\name{upgrade_internal}
\alias{upgrade_internal}
\title{Helper function to actually do upgrade}
\usage{
upgrade_internal(object)
}
\description{
Helper function to actually do upgrade
}
\keyword{internal}

