% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{valid_names}
\alias{valid_names}
\title{Make Valid Names}
\usage{
valid_names(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector
}
\description{
If names are invalid, quotes them with backtics
}
\keyword{internal}
