% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/options.R
\name{validate_options}
\alias{validate_options}
\title{Checks that options meet expectations before anything gets run}
\usage{
validate_options(opts.to.validate)
}
\description{
Checks that options meet expectations before anything gets run
}

