% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/text.R
\name{word_cat}
\alias{word_cat}
\title{Print To Screen Wrapping Words}
\usage{
word_cat(..., sep = " ", width = getOption("width"), tolerance = 8L,
  file = stdout())
}
\description{
Print To Screen Wrapping Words
}
\seealso{
\code{\link{word_wrap}}
}
\keyword{internal}

