% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{append,factor,factor-method}
\alias{append,factor,factor-method}
\title{Append Factors}
\usage{
\S4method{append}{factor,factor}(x, values, after = length(x))
}
\arguments{
\item{x}{the vector to be modified.}

\item{values}{to be included in the modified vector.}

\item{after}{a subscript, after which the values are to be appended.}
}
\description{
Note this is not related to \code{`\link{append,unitizerList,ANY-method}`}
except in as much as it is the same generic, so it just got thrown in here.
}
\keyword{internal}

