% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{as.character,unitizerItemTestsErrors-method}
\alias{as.character,unitizerItemTestsErrors-method}
\title{Convert An Error Into Character}
\usage{
\S4method{as.character}{unitizerItemTestsErrors}(x, ...)
}
\description{
Convert An Error Into Character
}
\keyword{internal}

