% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciiml.R
\name{as.character.bullet}
\alias{as.character.bullet}
\title{Produce Character Vector Representation of Bullet Lists}
\usage{
\method{as.character}{bullet}(x, width = 0L, ...)
}
\arguments{
\item{x}{object to render}

\item{width}{how many characters to wrap at}

\item{...}{dots, other arguments to pass to \code{word_wrap}}

\item{pre}{what to pre-pend to each bullet}
}
\value{
character vector containing rendered object, where each element
  corresponds to a line
}
\description{
Produce Character Vector Representation of Bullet Lists
}
\keyword{internal}
