% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\docType{methods}
\name{as.expression,unitizerList-method}
\alias{as.expression,unitizerList-method}
\title{Coerce to expression by returning items coerced to expressions}
\usage{
\S4method{as.expression}{unitizerList}(x, ...)
}
\description{
Really only meaningful for classes that implement the \code{.items}
slot as an expression, but works for others to the extent
\code{.items} contents are coercible to expressions
}
\keyword{internal}
