% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/list.R
\docType{methods}
\name{as.list,unitizerList-method}
\alias{as.list,unitizerList-method}
\title{Coerce to list by returning items}
\usage{
\S4method{as.list}{unitizerList}(x, ...)
}
\description{
Coerce to list by returning items
}
\keyword{internal}

