% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{c,unitizerList-method}
\alias{c,unitizerList-method}
\title{Concatenate to a \code{`\link{unitizerList-class}`}}
\usage{
\S4method{c}{unitizerList}(x, ..., recursive = FALSE)
}
\arguments{
\item{...}{objects to be concatenated.}

\item{recursive}{logical.  If \code{recursive = TRUE}, the function
    recursively descends through lists (and pairlists) combining all
    their elements into a vector.}
}
\description{
Concatenate to a \code{`\link{unitizerList-class}`}
}
\keyword{internal}

