% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{cap_first}
\alias{cap_first}
\alias{decap_first}
\title{Captalizes or Decapitalizes First Letter}
\usage{
cap_first(x)
}
\arguments{
\item{x}{character}
}
\value{
character
}
\description{
Captalizes or Decapitalizes First Letter
}
\keyword{internal}
