% Generated by roxygen2 (4.0.2): do not edit by hand
\name{comm_extract}
\alias{comm_extract}
\title{Utility Function to Extract Comments From Expression}
\usage{
comm_extract(x)
}
\description{
Note that when dealing with expressions the very first item will typically
be NULL to allow for logic that works with nested structures
}
\keyword{internal}

