% Generated by roxygen2 (4.0.2): do not edit by hand
\name{deparse_call}
\alias{deparse_call}
\title{Used to generate character values to store in cached deparse list}
\usage{
deparse_call(expr)
}
\arguments{
\item{expr}{language to deparse}
}
\value{
character(1L)
}
\description{
Used to generate character values to store in cached deparse list
}
\keyword{internal}

