% Generated by roxygen2 (4.0.2): do not edit by hand
\name{deparse_prompt}
\alias{deparse_prompt}
\title{Deparse, But Make It Look Like It Would On Prompt}
\usage{
deparse_prompt(expr)
}
\arguments{
\item{expr}{an expression or call}
}
\value{
character vector
}
\description{
Deparse, But Make It Look Like It Would On Prompt
}

