% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/browse.struct.R
\name{deparseCalls}
\alias{deparseCalls}
\title{Pull Out Deparsed Calls From Objects}
\usage{
deparseCalls(x, ...)
}
\value{
character the deparsed calls
}
\description{
Used primarily as a debugging tool, should probably be migrated to use
\code{`\link{extractItems}`}
}
\keyword{internal}

