% Generated by roxygen2 (4.0.1): do not edit by hand
\name{desc}
\alias{desc}
\title{Produces 1 Line Description of Value}
\usage{
desc(val, limit = getOption("width"))
}
\arguments{
\item{val}{object to describe}

\item{limit}{max characters to display}
}
\value{
character vector describing object
}
\description{
Produces 1 Line Description of Value
}
\keyword{internal}

