% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{desc}
\alias{desc}
\title{One Line Description of Object}
\usage{
desc(val, depth = 0, limit = getOption("width"))
}
\arguments{
\item{val}{object to describe}

\item{limit}{max characters to display}
}
\value{
character(1L) describing object
}
\description{
Objects are described by class, and dimensions.  Dimensions is always denoted
in square brackets.  For example, \dQuote{int[10]} means an integer of length
ten.
}
\examples{
desc(list(a=iris, b=lm(dist ~ speed, cars), 1:10, matrix(letters, 2))
}

