% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/text.R
\name{diff_obj_out}
\alias{diff_obj_out}
\alias{obj_capt}
\alias{obj_screen_out}
\title{Show a Faux Diff Between Two Objects}
\usage{
diff_obj_out(obj.rem, obj.add,
  obj.rem.name = deparse(substitute(obj.rem))[[1L]],
  obj.add.name = deparse(substitute(obj.add))[[1L]],
  width = getOption("width"), max.len = NULL, file = stdout(),
  frame = parent.frame())
}
\arguments{
\item{obj.rem}{object to compare}

\item{obj.add}{object to compare}

\item{obj.rem.name}{object to compare}

\item{obj.add.name}{object to compare}

\item{width}{at what width to wrap output}

\item{max.len}{2 length integer vector with first value threshold at which we
start trimming output and the second the length we tri to}

\item{file}{whether to show to stdout or stderr}

\item{frame}{what frame to capture in, relevant mostly if looking for a print
method}

\item{x}{a character vector}

\item{y}{another character vector to compare to}
}
\value{
\itemize{
     \item for \code{char_diff} a list with two vectors, each of same length
       as inputs, where FALSE indicates value is the same as in the other
       vector, and TRUE indicates it is different
  }
}
\description{
The idea is to save the user the need to print out the objects to screen.
This will print two objects to screen in a faux git diff look, focusing on a
snippet of the object.
}
\keyword{internal}

