% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{done,unitizerBrowse-method}
\alias{done,unitizerBrowse-method}
\title{Indicate Whether to Exit Review Loop}
\usage{
\S4method{done}{unitizerBrowse}(x, ...)
}
\description{
Indicate Whether to Exit Review Loop
}
\keyword{internal}

