% Generated by roxygen2 (4.0.2): do not edit by hand
\name{env_ancestry}
\alias{env_ancestry}
\title{Retrieves Environment Ancestry}
\usage{
env_ancestry(env, stop.env = globalenv())
}
\arguments{
\item{env}{the environment to start with}

\item{stop.env}{the environment to stop with}
}
\description{
Retrieves Environment Ancestry
}
\keyword{internal}

