% Generated by roxygen2 (4.0.2): do not edit by hand
\name{header_help}
\alias{header_help}
\title{Helper function for single line headers}
\usage{
header_help(x, ..., pad.char = "-")
}
\arguments{
\item{x}{the contents of the header}

\item{...}{unused, for compatibility with print generic}

\item{pad.char}{which character to use to form the header structure}
}
\description{
Helper function for single line headers
}
\keyword{internal}

