% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/asciiml.R
\name{header_help}
\alias{header_help}
\title{Helper function for single line headers}
\usage{
header_help(x, width, ..., pad.char = "-")
}
\arguments{
\item{x}{the contents of the header}

\item{width}{how wide we want the header to display}

\item{...}{unused, for compatibility with print generic}

\item{pad.char}{which character to use to form the header structure}
}
\description{
Helper function for single line headers
}
\keyword{internal}

