% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/item.R
\docType{methods}
\name{initialize,unitizerItem-method}
\alias{initialize,unitizerItem-method}
\title{Initialize A \code{`\link{unitizerItem-class}`}}
\usage{
\S4method{initialize}{unitizerItem}(.Object, ...)
}
\description{
Makes the fact that most of the data needs to be part of a
\code{`\link{unitizerItemData-class}`} object transparent to the user.
}
\keyword{internal}

