% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{initialize,unitizerItem-method}
\alias{initialize,unitizerItem-method}
\title{Initialize A \code{`\link{unitizerItem-class}`}}
\usage{
\S4method{initialize}{unitizerItem}(.Object, ...)
}
\arguments{
\item{.Object}{ An object:  see the Details section.}

\item{...}{data to include in the new object.  Named arguments
    correspond to slots in the class definition. Unnamed arguments must
    be objects from classes that this class extends.}
}
\description{
Makes the fact that most of the data needs to be part of a
\code{`\link{unitizerItemData-class}`} object transparent to the user.
}
\keyword{internal}

