% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{length,unitizer-method}
\alias{length,unitizer-method}
\title{Compute Length of a \code{`\link{unitizer-class}`} Object}
\usage{
\S4method{length}{unitizer}(x)
}
\arguments{
\item{x}{an \R object.  For replacement, a vector or factor.}
}
\description{
Compute Length of a \code{`\link{unitizer-class}`} Object
}
\keyword{internal}

