% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unitizer.R
\docType{methods}
\name{length,unitizer-method}
\alias{length,unitizer-method}
\title{Compute Length of a \code{\link{unitizer-class}} Object}
\usage{
\S4method{length}{unitizer}(x)
}
\description{
Compute Length of a \code{\link{unitizer-class}} Object
}
\keyword{internal}

