% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/browse.struct.R
\docType{methods}
\name{length,unitizerBrowseSubSection-method}
\alias{length,unitizerBrowseSubSection-method}
\title{Compute Length of a \code{`\link{unitizerBrowseSubSection-class}`}}
\usage{
\S4method{length}{unitizerBrowseSubSection}(x)
}
\description{
Compute Length of a \code{`\link{unitizerBrowseSubSection-class}`}
}
\keyword{internal}

