% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{length,unitizerBrowseSubSection-method}
\alias{length,unitizerBrowseSubSection-method}
\title{Compute Length of a \code{`\link{unitizerBrowseSubSection-class}`}}
\usage{
\S4method{length}{unitizerBrowseSubSection}(x)
}
\arguments{
\item{x}{an \R object.  For replacement, a vector or factor.}
}
\description{
Compute Length of a \code{`\link{unitizerBrowseSubSection-class}`}
}
\keyword{internal}

