% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/change.R
\docType{methods}
\name{length,unitizerChanges-method}
\alias{length,unitizerChanges-method}
\title{Return Sum of Total Changes}
\usage{
\S4method{length}{unitizerChanges}(x)
}
\description{
Return Sum of Total Changes
}
\keyword{internal}

