% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/list.R
\docType{methods}
\name{length,unitizerList-method}
\alias{length,unitizerList-method}
\title{Compute Number of Items in \code{`\link{unitizerList-class}`}}
\usage{
\S4method{length}{unitizerList}(x)
}
\description{
Compute Number of Items in \code{`\link{unitizerList-class}`}
}
\keyword{internal}

