% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{length,unitizerSection-method}
\alias{length,unitizerSection-method}
\title{Compute Length of a \code{unitizerSection-class}}
\usage{
\S4method{length}{unitizerSection}(x)
}
\arguments{
\item{x}{a \code{unitizerSection} object}
}
\description{
Compute Length of a \code{unitizerSection-class}
}
\keyword{internal}
