% Generated by roxygen2 (4.0.2): do not edit by hand
\name{navigate_prompt}
\alias{navigate_prompt}
\title{Wrapper Around User Interaction}
\usage{
navigate_prompt(x, curr.id, text, browse.env1 = globalenv(),
  browse.env2 = globalenv(), help = character(), valid.opts)
}
\arguments{
\item{x}{a unitizer.browse object}

\item{curr.id}{which id we are currently browsing}

\item{text}{the prompt text to display}

\item{browse.env1}{environment to have user review tests, run commands, etc}

\item{browse.env2}{navigation environment}

\item{help}{a character vector with help suggestions}
}
\description{
Specifically for cases were user has the choice to input something or to try
to navigate to another test.
}
\seealso{
unitizer_prompt
}
\keyword{internal}

