% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/text.R
\name{one_line}
\alias{one_line}
\title{Collapse Multi-line Character into one line}
\usage{
one_line(x, chars = 0L)
}
\arguments{
\item{x}{character}

\item{chars}{how many characters to display}
}
\description{
Collapse Multi-line Character into one line
}
\keyword{internal}

