% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unitizer.R
\name{passed}
\alias{passed}
\title{Determine if a \code{unitizer} Passed Based On Summary}
\usage{
passed(object, ...)
}
\arguments{
\item{object}{object to test for passing}
}
\value{
logical(1L)
}
\description{
Determine if a \code{unitizer} Passed Based On Summary
}
\keyword{internal}

