% Generated by roxygen2 (4.0.2): do not edit by hand
\name{path_clean}
\alias{path_clean}
\title{Cleans a Path to be In Standard Format}
\usage{
path_clean(path)
}
\arguments{
\item{path}{character the path name to clean up}
}
\value{
the cleaned up path
}
\description{
Uses \code{`\link{dirname}`} to convert paths on windows machines with back
slasshes to forward slash based names, and then removed excess forward
slashes.
}
\keyword{internal}

