\docType{methods}
\name{+,testor,testorItem-method}
\alias{+,testor,testorItem-method}
\title{Adds \code{`\link{testorItem}`} to \code{`\link{testor}`}}
\usage{
\S4method{+}{testor,testorItem}(e1, e2)
}
\description{
All tests are run on addition, and mapping information
between reference and new tests is also recored.
}
\seealso{
\code{`\link{registerItem,testor,testorItem-method}`}
}
\keyword{internal}

