% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{+,unitizer,unitizerItems-method}
\alias{+,unitizer,unitizerItems-method}
\title{Adds \code{`\link{unitizerItems-class}`} objects to unitizer}
\usage{
\S4method{+}{unitizer,unitizerItems}(e1, e2)
}
\description{
Any added \code{`\link{unitizerItems-class}`} objects are treated as
reference items.  The only way to add new items is by adding each
item individually with \code{`\link{+,unitizer,unitizerItem-method}`}.
}
\details{
One aspect of copying reference items which isn't handled here is moving
over the section data because this is kept at the \code{`\link{unitizer-class}`}
level, not at the \code{`\link{unitizerItems-class}`} level.  The
section copying is handled by \code{`\link{refSections,unitizer,unitizer-method}`}.
This is something that we should clean-up eventually.
}
\keyword{internal}

