% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{+,unitizer,unitizerItemTestsErrors-method}
\alias{+,unitizer,unitizerItemTestsErrors-method}
\title{Add Test Errors to \code{`\link{unitizer-class}`}}
\usage{
\S4method{+}{unitizer,unitizerItemTestsErrors}(e1, e2)
}
\description{
Add Test Errors to \code{`\link{unitizer-class}`}
}
\keyword{internal}

