% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unitizer.add.R
\docType{methods}
\name{+,unitizer,unitizerSection-method}
\alias{+,unitizer,unitizerSection-method}
\title{Add a \code{`\link{unitizerSection-class}`} to a \code{`\link{unitizer-class}`}}
\usage{
\S4method{+}{unitizer,unitizerSection}(e1, e2)
}
\description{
Registers the section, and the mapping of items to section.
}
\keyword{internal}

