% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{+,unitizerBrowse,unitizerBrowseSection-method}
\alias{+,unitizerBrowse,unitizerBrowseSection-method}
\title{Add Sections to Our Main Browse Object}
\usage{
\S4method{+}{unitizerBrowse,unitizerBrowseSection}(e1, e2)
}
\description{
Primarily we're contructing the \code{`@mapping`} slot which will then allow
us to carry out requisite computations later.  See \code{`\link{unitizerBrowseMapping-class}`}
for details on what each of the slots in \code{`mapping`} does.
}
\details{
Also, some more discussion of this issue in the docs for \code{`\link{unitizer-class}`}.
}
\keyword{internal}

