% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{+,unitizerBrowseSection,unitizerBrowseSubSection-method}
\alias{+,unitizerBrowseSection,unitizerBrowseSubSection-method}
\title{Add a browsing sub-section to a browse section}
\usage{
\S4method{+}{unitizerBrowseSection,unitizerBrowseSubSection}(e1, e2)
}
\arguments{
\item{e1}{a \code{`\link{unitizerBrowseSection-class}`}}

\item{e2}{a \code{`\link{unitizerBrowseSubSection-class}`}}
}
\value{
a \code{`\link{unitizerBrowseSection-class}`}
}
\description{
Add a browsing sub-section to a browse section
}
\keyword{internal}

