% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{+,unitizerItems,unitizerItems-method}
\alias{+,unitizerItems,unitizerItems-method}
\title{Add a \code{`\link{unitizerItem-class}`} to a \code{`\link{unitizerItems-class}`}}
\usage{
\S4method{+}{unitizerItems,unitizerItems}(e1, e2)
}
\description{
Add a \code{`\link{unitizerItem-class}`} to a \code{`\link{unitizerItems-class}`}
}
\keyword{internal}

