% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repairenvs.R
\name{repair_environments}
\alias{repair_environments}
\title{Repair Environment Chains}
\usage{
repair_environments(x, interactive.mode = interactive())
}
\arguments{
\item{x}{either a unitizer, or a store id (see \code{\link{unitize}})}

\item{interactive.mode}{logical(1L) whether to run in interactive mode (
request user input when needed) or not (error if user input is required,
e.g. if all tests do not pass).}
}
\value{
a \code{unitizer} object
}
\description{
In theory should never be needed, but use in case you get errors about
corrupted environments.  You should only use this if you get an error
telling you to use it.
}
\details{
If you pass a store id this will re-save the repaired \code{unitizer} to
the location specified by the store id.
}
\seealso{
\code{\link{unitize}}
}
