% Generated by roxygen2 (4.0.2): do not edit by hand
\name{review_prompt}
\alias{review_prompt}
\title{Manages Producing Test Navigation Message / Prompt}
\usage{
review_prompt(x, nav.env)
}
\arguments{
\item{x}{a unitizerBrowse object}

\item{nav.env}{an environment}
}
\value{
either a \code{`unitizerBrowse`}, or "Q" if the user chose to quit
}
\description{
Probably should be an S4 method, along with \code{`\link{navigate_prompt}`}
}
\keyword{internal}

