% Generated by roxygen2 (4.0.2): do not edit by hand
\name{search_path_check}
\alias{search_path_check}
\title{Reconstruct Search Path From History}
\usage{
search_path_check(verbose = FALSE)
}
\arguments{
\item{verbose}{whether to output details of failures, purely for internal debugging}
}
\description{
This is an internal check to make sure the shims on \code{`library/require/attach/detach`}
worked correctly.
}
\keyword{internal}

