% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{show,conditionList-method}
\alias{show,conditionList-method}
\title{Prints A list of Conditions}
\usage{
\S4method{show}{conditionList}(object)
}
\arguments{
\item{object}{a \code{`conditionList`} object (list of conditions)}
}
\value{
object, invisibly
}
\description{
Prints A list of Conditions
}
\keyword{internal}

