% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\docType{methods}
\name{show,unitizerChanges-method}
\alias{show,unitizerChanges-method}
\title{Print Out A Summary Of the Changes}
\usage{
\S4method{show}{unitizerChanges}(object)
}
\description{
Print Out A Summary Of the Changes
}
\keyword{internal}
