% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/item.sub.R
\docType{methods}
\name{show,unitizerItemTestsErrors-method}
\alias{show,unitizerItemTestsErrors-method}
\title{Display Test Errors}
\usage{
\S4method{show}{unitizerItemTestsErrors}(object)
}
\description{
Display Test Errors
}
\keyword{internal}

