% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show,unitizerSummary-method}
\alias{show,unitizerSummary-method}
\title{Display Unitizer Summary}
\usage{
\S4method{show}{unitizerSummary}(object)
}
\arguments{
\item{object}{the object to show}
}
\description{
Unfortunately no choice but to use \code{`getOptions("width")`} from within
here.  Maybe could pre-compute in one of earlier stages and stuff into
\code{`object`}?  Not a big deal
}
\keyword{internal}

