% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{show.conditionList}
\alias{show.conditionList}
\alias{show,conditionList-method}
\title{Prints A list of Conditions}
\usage{
\S4method{show}{conditionList}(object)
}
\arguments{
\item{object}{a \code{\link{conditionList}} object (list of conditions)}
}
\value{
object, invisibly
}
\description{
S4 method for \code{\link{conditionList}} objects.
}
\examples{
## Create a test item as you would find normally at the `unitizer` prompt
## for illustrative purposes:
.NEW <- mock_item()
## Show the conditions the test generated (typing `show` here is optional
## since auto-printing should dispatch to `show`)
show(.NEW$conditions)
}
\seealso{
\code{\link{conditionList}}
}
