% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prompt.R
\name{simple_prompt}
\alias{simple_prompt}
\title{A Simple Prompting Function}
\usage{
simple_prompt(message, values = c("Y", "N"), prompt = "unitizer> ",
  attempts = 3L, case.sensitive = FALSE)
}
\arguments{
\item{message}{character ask the user a question}

\item{values}{character valid responses}

\item{prompt}{see \code{\link{readline}}}

\item{attempts}{how many times to try before erroring}
}
\value{
one of \code{values} as selected by user
}
\description{
A Simple Prompting Function
}

