% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{[,unitizerList,subIndex,missing,missing-method}
\alias{[,unitizerList,subIndex,missing,missing-method}
\title{Subsetting Method for \code{`\link{unitizerList-class}`}}
\usage{
\S4method{[}{unitizerList,subIndex,missing,missing}(x, i)
}
\description{
Subsetting Method for \code{`\link{unitizerList-class}`}
}
\keyword{internal}

