\docType{methods}
\name{summary,testor-method}
\alias{summary,testor-method}
\title{Summarize Results}
\usage{
\S4method{summary}{testor}(object, ...)
}
\arguments{
  \item{object}{an object for which a summary is desired.}

  \item{...}{additional arguments affecting the summary
  produced.}
}
\description{
Summarize Results
}
\keyword{internal}

