% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unitizer.R
\docType{methods}
\name{summary,unitizer-method}
\alias{summary,unitizer-method}
\title{Summarize Results}
\usage{
\S4method{summary}{unitizer}(object, ...)
}
\value{
a list with the data that can be assigned to sections in vector/matrix
  form, and the section less stuff (Deletes) as a scalar
}
\description{
Summarize Results
}
\keyword{internal}

