% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unitizer.R
\docType{methods}
\name{summary,unitizer-method}
\alias{summary,unitizer-method}
\title{Summarize Results}
\usage{
\S4method{summary}{unitizer}(object, silent = FALSE, ...)
}
\arguments{
\item{object}{the object to summarize}

\item{silent}{whether to suppress display of summary object}
}
\value{
a unitizerSummary object
}
\description{
Also prints to screen, but only if \code{level == 1L}
}
\keyword{internal}

