% Generated by roxygen2 (4.0.1): do not edit by hand
\name{unitize}
\alias{unitize}
\title{Unitize an R Test Script}
\usage{
unitize(test.file, store.id = sub("\\\\.[Rr]$", ".unitizer", test.file),
  interactive.mode = interactive())
}
\arguments{
\item{test.file}{path to the file containing tests}

\item{store.id}{a folder to store the \code{`unitizer`} objects in; will auto-
generate to a folder at the same location as the test file with the same
name as the testfile, except ending in \code{`.unitizer`} instead of \code{`.R`}}

\item{interactive.mode}{logical(1L) whether to run in interactive mode}
}
\description{
Turns a standard R script into unit tests by evaluating the expressions and
storing them along with their resuls.  Re-running \code{`unitize`} then
checks that the values remain unchanged.  See vignette for more details.
}
\details{
You can run \code{`unitize`} from the command line, or you can place one or
more \code{`unitize`} calls in an R file.
}
\seealso{
\code{`\link{get_store}`}
}

