% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitizer-package.r
\docType{package}
\name{unitizer}
\alias{unitizer}
\title{unitizer}
\description{
Simplifies regression tests by comparing objects produced by test
code with earlier versions of those same objects.  If objects are unchanged
the tests pass.  `unitizer` provides an interactive interface to review
failing tests or new tests.  See vignettes for details.
}
