% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls.R
\name{unitizer_ls}
\alias{unitizer_ls}
\title{An `ls` Like Function}
\usage{
unitizer_ls(name, pos = -1L, envir = parent.frame(),
  all.names = FALSE, pattern)
}
\value{
list of object names, or a list with environments containing the
  objects
}
\description{
Much like `ls`, except that it is designed to crawl up the \code{`.new`} and
\code{`.ref`} environments and display all the objects.
}
\details{
This is used in \code{`browseUnitizer,unitizer-unitizerBrowse-method`},
and is re-assigned to \code{`ls`} for use in the \code{`unitizer`} prompt
environment.
}
\keyword{internal}
