% Generated by roxygen2 (4.0.2): do not edit by hand
\name{unitizer_quit}
\alias{unitizer_quit}
\alias{unitizer_quit_handler}
\title{Overrides Default quit() Behavior}
\usage{
unitizer_quit(save = "default", status = 0, runLast = TRUE)
}
\description{
Necessary because quit short circuits the `on.exit` clean-up functions and
would leave stuff in a weird state (history not reset, etc.).
}
\details{
This is used in \code{`\link{unitize}`}.
}
\keyword{internal}

