% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\docType{class}
\name{unitizerChanges-class}
\alias{unitizerChanges-class}
\title{Summary of Changes}
\description{
Changes arise any time a user, through the interactive unitizer mode,
requests the storing of a change (accept new version of failed test,
add new test, remove old test).
}
\keyword{internal}
