% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{unitizerItemTestError-class}
\alias{unitizerItemTestError-class}
\alias{unitizerItemTestsErrors-class}
\alias{unitizerItemsTestsErrors-class,}
\title{Stores Errors from Evaluating New vs. Ref Comparisons}
\description{
There various nested objects involved:
\itemize{
  \item \code{`unitizerItemTestError`} contains the error produced from a comparison
  \item \code{`unitizerItemTestsErrors`} aggregates the errors for each slot of an item
  \item \code{`unitizerItemsTestsErrors`} aggregates all the errors for the
     \code{`\link{unitizer-class}`} object
}
}
\keyword{internal}

