% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{unitizerItemTestFun-class}
\alias{unitizerItemTestFun-class}
\title{Validates Functions for Use in New vs. Reference Object Comparison}
\description{
Validates Functions for Use in New vs. Reference Object Comparison
}
\keyword{internal}

