% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\docType{class}
\name{unitizerSection-class}
\alias{unitizerSection-class}
\alias{unitizerSectionExpression-class}
\alias{unitizerSectionNA-class}
\title{Contains Representation For a Section of Tests}
\description{
\code{unitizerSectionExpression-class} contains the actual expressions that
belong to the section, whereas \code{unitizerSection-class} only contains
the meta data.  The latter objects are used within \code{]unitizer-class},
whereas the former is really just a temporary object until we can generate
the latter.
}
\details{
\code{unitizerSectionNA-class} is a specialized section for tests that actually
don't have a section (removed tests that are nonetheless chosen to be kept
by user in interactive environment)
}
\section{Slots}{

\describe{
\item{\code{title}}{1 lenght character, the name of the section}

\item{\code{details}}{character vector containing additional info on the section}

\item{\code{compare}}{functions to compare the various aspects of a
\code{unitizerItem-class} @slot length tracks size of the section}
}}

\keyword{internal}
