% Generated by roxygen2 (4.0.2): do not edit by hand
\name{word_cat}
\alias{word_cat}
\title{Cat but split by words to allow wrapping}
\usage{
word_cat(..., fill = TRUE)
}
\description{
Parameters are the same as \code{`\link{cat}`}.  Word splitting and auto-fill
only happens if the \code{`sep`} argument is not provided.
}
\seealso{
\code{`\link{cat}`}
}
\keyword{internal}

